% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{addDataset,PipelineTemplate-method}
\alias{addDataset,PipelineTemplate-method}
\title{Set the addDataset slot value for the object}
\usage{
\S4method{addDataset}{PipelineTemplate}(obj, name, dataset)
}
\arguments{
\item{obj}{xxxx}

\item{name}{xxx}

\item{dataset}{xxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
