% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggregateIter}
\alias{aggregateIter}
\title{xxxx}
\usage{
aggregateIter(obj.pep, X, init.method = "sum", method = "mean", ...)
}
\arguments{
\item{obj.pep}{xxxxx}

\item{X}{xxxx}

\item{init.method}{xxxxx}

\item{method}{xxxxx}

\item{...}{xxxx}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\author{
Samuel Wieczorek
}
