% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{AggregateMetacell}
\alias{AggregateMetacell}
\title{Symbolic product of matrices}
\usage{
AggregateMetacell(mat, obj.pep)
}
\arguments{
\item{mat}{An adjacency matrix between peptides and proteins}

\item{obj.pep}{A dataframe of the cell metadata for peptides}
}
\value{
xxxx
}
\description{
Execute a product two matrices: the first is an adjacency one  while the
second if a simple dataframe
}
\author{
Samuel Wieczorek
}
