% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{boxPlotD}
\alias{boxPlotD}
\title{Builds a boxplot from a dataframe}
\usage{
boxPlotD(qData, dataForXAxis = NULL, conds = NULL, palette = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{dataForXAxis}{A vector containing the types of replicates 
to use as X-axis. Available values are: Condition, Bio.Rep,
Bio.Rep and Tech.Rep. Default is "Condition".}

\item{conds}{A vector of the conditions (one string per sample).}

\item{palette}{xxx}
}
\value{
A boxplot
}
\description{
Boxplot for quantitative proteomics data
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
types <- c("Condition","Bio.Rep")
dataForXAxis <- Biobase::pData(Exp1_R25_pept)[,types]
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
boxPlotD(qData, dataForXAxis, conds)
}
\seealso{
\code{\link{densityPlotD}}
}
\author{
Florence Combes, Samuel Wieczorek
}
