% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{BuildColumnToProteinDataset_par}
\alias{BuildColumnToProteinDataset_par}
\title{creates a column for the protein dataset after agregation by
 using the previous peptide dataset.}
\usage{
BuildColumnToProteinDataset_par(peptideData, matAdj, columnName,
  proteinNames)
}
\arguments{
\item{peptideData}{A data.frame of meta data of peptides. It is the fData 
of the MSnset object.}

\item{matAdj}{The adjacency matrix used to agregate the peptides data.}

\item{columnName}{The name of the column in fData(peptides_MSnset) that 
the user wants to keep in the new protein data.frame.}

\item{proteinNames}{The names of the protein in the new dataset (i.e. rownames)}
}
\value{
A vector
}
\description{
This function creates a column for the protein dataset after agregation 
by using the previous peptide dataset. It is a parallel version of the function
\code{BuildColumnToProteinDataset}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
protID <- "Protein.group.IDs"
M <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], protID, FALSE)
data <- Biobase::fData(Exp1_R25_pept[1:1000])
protData <- pepAgregate(Exp1_R25_pept[1:1000], ProtID, 'sum overall', M)
name <- "Protein.group.IDs"
proteinNames <- rownames(Biobase::fData(protData))
BuildColumnToProteinDataset_par(data, M, name,proteinNames )
}
\author{
Samuel Wieczorek
}
