% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{buildLogText}
\alias{buildLogText}
\title{Build the text information to be saved}
\usage{
buildLogText(name, l.params)
}
\arguments{
\item{name}{The name of the process in Prostar}

\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
Build the text information to be saved after a process on an object of class \code{MSnSet}
}
\examples{
buildLogText("Original", list(filename="foo.MSnset"))
}
\author{
Samuel Wieczorek
}
