% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{BuildMetaCell}
\alias{BuildMetaCell}
\title{xxxx}
\usage{
BuildMetaCell(from = NULL, qData = NULL, conds = NULL, df = NULL)
}
\arguments{
\item{from}{xxx}

\item{qData}{An object of class \code{MSnSet}}

\item{conds}{xxx}

\item{df}{A list of integer xxxxxxx}
}
\value{
xxxxx
}
\description{
xxxxxx
}
\examples{
file <- system.file("extdata", "Exp1_R25_pept.txt", package="DAPARdata")
data <- read.table(file, header=TRUE, sep="\t",stringsAsFactors = FALSE)
metadataFile <- system.file("extdata", "samples_Exp1_R25.txt", package="DAPARdata")
metadata <- read.table(metadataFile, header=TRUE, sep="\t", as.is=TRUE, stringsAsFactors = FALSE)
conds <- metadata$Condition
qData <- data[,56:61]
df <- data[ , 43:48]
df <- BuildMetaCell(from = 'maxquant', qData = qData, conds = conds, df = df)
df <- BuildMetaCell(from = 'proline', qData = qData, conds = conds, df = df)

}
\author{
Samuel Wieczorek
}
