% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BuildPalette}
\alias{BuildPalette}
\title{djlqshdhsq dq idqs dqsd.}
\usage{
BuildPalette(conds, base_palette)
}
\arguments{
\item{conds}{xxxxx.}

\item{base_palette}{xxxxx;}
}
\value{
xxxxxx
}
\description{
sfdsfdsfdsfd
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
getProcessingInfo(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
