% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-pipeline-class.R
\name{CC,PipelinePeptide-method}
\alias{CC,PipelinePeptide-method}
\title{sdfdsfs}
\usage{
\S4method{CC}{PipelinePeptide}(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
