% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{check.conditions}
\alias{check.conditions}
\title{Check if the design is valid}
\usage{
check.conditions(conds)
}
\arguments{
\item{conds}{A vector}
}
\value{
A list
}
\description{
Check if the design is valid
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
check.conditions(Biobase::pData(Exp1_R25_pept)$Condition)

}
\author{
Samuel Wieczorek
}
