% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{check.design}
\alias{check.design}
\title{Check if the design is valid}
\usage{
check.design(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the `pData()` function 
of package `MSnbase`.}
}
\value{
A boolean
}
\description{
Check if the design is valid
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
check.design(Biobase::pData(Exp1_R25_pept)[, seq_len(3)])

}
\author{
Thomas Burger, Samuel Wieczorek
}
