% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{check.replicates.fullfilled}
\alias{check.replicates.fullfilled}
\title{Check if the design is valid}
\usage{
check.replicates.fullfilled(res = NULL, sTab)
}
\arguments{
\item{res}{A list}

\item{sTab}{a data.frame xxxx}
}
\value{
A list
}
\description{
Check if the design is valid
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
sTab <- Biobase::pData(Exp1_R25_pept)
check.replicates.fullfilled(sTab = sTab)

}
\author{
Samuel Wieczorek
}
