% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{checkClusterability}
\alias{checkClusterability}
\title{Prepare the data for subsequent clustering}
\usage{
checkClusterability(standards, b = 500)
}
\arguments{
\item{standards}{a dataframe containing the standardized mean intensities}
}
\value{
a list of 2 elements:
* dip_test: the result of the clusterability of the data
* gap_cluster: the gap statistic obtained with the function [cluster::clusGap()].
}
\description{
The first step is to standardize the data (with the \code{Mfuzz}
package). Then the function checks that these data are clusterizable or not
 (use of [diptest::dip.test()] to determine whether the distribution is unimodal or
 multimodal). Finally, it determines the "optimal" k by the Gap statistic
 approach.
}
\examples{
utils::data(Exp1_R25_prot, package='DAPARdata')
obj <- Exp1_R25_prot[1:1000]
keepThat <- mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
st <- standardiseMeanIntensities(obj)
checkClusterability(st, b=100)

}
\author{
Hélène Borges
}
