% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{classic1wayAnova}
\alias{classic1wayAnova}
\title{Function to perform a One-way Anova statistical test on a MsnBase dataset}
\usage{
classic1wayAnova(current_line, conditions)
}
\arguments{
\item{current_line}{The line currently treated from the quantitative data to
perform the ANOVA}

\item{conditions}{The conditions represent the different classes of the
studied factor}
}
\value{
A named vector containing all the different values of the aov model
}
\description{
Function to perform a One-way Anova statistical test on a MsnBase dataset
}
\author{
Hélène Borges
}
