% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{compareNormalizationD}
\alias{compareNormalizationD}
\title{Builds a plot from a dataframe}
\usage{
compareNormalizationD(qDataBefore, qDataAfter, labelsForLegend = NULL,
  indData2Show = NULL, group2Color = "Condition")
}
\arguments{
\item{qDataBefore}{A dataframe that contains quantitative data before normalization.}

\item{qDataAfter}{A dataframe that contains quantitative data after normalization.}

\item{labelsForLegend}{A vector of the conditions (labels) (one label per sample).}

\item{indData2Show}{A vector of labels to show in xxx}

\item{group2Color}{xxxxxx}
}
\value{
A plot
}
\description{
Plot to compare the quantitative proteomics data before and after normalization
}
\examples{
data(UPSprotx2)
qDataBefore <- exprs(UPSprotx2)
labels <- pData(UPSprotx2)[,"Label"]
qDataAfter <- normalizeD(qDataBefore, labels, "Median Centering", "within conditions")
compareNormalizationD(qDataBefore, qDataAfter, labels)
}
\author{
Samuel Wieczorek
}

