% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{compute.selection.table}
\alias{compute.selection.table}
\title{xxx}
\usage{
compute.selection.table(x, fdr.threshold)
}
\arguments{
\item{x}{xxx}

\item{fdr.threshold}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
NULL

}
\author{
Thomas Burger
}
