% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{corrMatrixD}
\alias{corrMatrixD}
\title{Displays a correlation matrix of the quantitative data of the
\code{exprs()} table.}
\usage{
corrMatrixD(obj, indLegend = NULL)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{indLegend}{A vector of indices in the columns in
the \code{pData()} table choosen for the labels in the axes.}
}
\value{
A colored correlation matrix
}
\description{
Correlation matrix based on a \code{\link{MSnSet}} object
}
\examples{
data(UPSprotx2)
corrMatrixD(UPSprotx2)
}
\author{
Florence Combes, Samuel Wieczorek
}

