% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_corr_matrix.R
\name{corrMatrixD_HC}
\alias{corrMatrixD_HC}
\title{Displays a correlation matrix of the quantitative data of the
\code{exprs()} table.}
\usage{
corrMatrixD_HC(object, samplesData = NULL, rate = 0.5, showDataLabels = TRUE)
}
\arguments{
\item{object}{The result of the \code{cor} function.}

\item{samplesData}{A dataframe in which lines correspond to samples and 
columns to the meta-data for those samples.}

\item{rate}{The rate parameter to control the exponential law for 
the gradient of colors}

\item{showDataLabels}{xxx}
}
\value{
A colored correlation matrix
}
\description{
#' Correlation matrix based on a \code{MSnSet} object
#' 
#' @title Displays a correlation matrix of the quantitative data of the
#' \code{exprs()} table.
#' 
#' @param qData A dataframe of quantitative data.
#' 
#' @param samplesData A dataframe where lines correspond to samples and 
#' columns to the meta-data for those samples.
#' 
#' @param gradientRate The rate parameter to control the exponential law for 
#' the gradient of colors
#' 
#' @return A colored correlation matrix
#' 
#' @author Florence Combes, Samuel Wieczorek
#' 
#' @examples
#' utils::data(Exp1_R25_pept, package='DAPARdata')
#' qData <- Biobase::exprs(Exp1_R25_pept)
#' samplesData <- Biobase::pData(Exp1_R25_pept)
#' corrMatrixD(qData, samplesData)
#' 
#' @export
#' 
#' @importFrom stats pexp cor
#' @importFrom ggplot2 element_text qplot theme labs scale_fill_gradientn
#' @importFrom reshape2 melt
#' 
corrMatrixD <- function(qData, samplesData, gradientRate = 5){
  Var1 <- Var2 <- value <- NULL
  
  for (j in 1:length(colnames(qData))){
    colnames(qData)[j] <- paste(as.character(samplesData[j,2:ncol(samplesData)]), 
                                collapse =" ")
  }
  
  z <- cor(qData,use = 'pairwise.complete.obs')
  text <- element_text(colour="black", size = 16, face = "bold")
  d <- qplot(x = Var1, 
             y = Var2, 
             data = melt(z), 
             fill = value, 
             geom = "tile") +
    theme(axis.text = element_text(size=16),
          axis.title = element_text(size=20, face="bold"),
          axis.text.x = element_text(angle=30, vjust=1, hjust=1),
          legend.text = text,
          legend.title = text) +
    labs(x = "", y = "") +
    
    scale_fill_gradientn (
      colours = colorRampPalette (c ("white", "lightblue","darkblue")) (101),
      values = c(pexp(seq(0,1,0.01), rate=gradientRate),1), limits=c(0,1))
  
  plot(d)
}
}
\details{
Correlation matrix based on a \code{MSnSet} object. Same as the 
function \link{corrMatrixD} but uses the package \code{highcharter}
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
qData <- Biobase::exprs(Exp1_R25_pept)
samplesData <- Biobase::pData(Exp1_R25_pept)
res <- cor(qData,use = 'pairwise.complete.obs')
corrMatrixD_HC(res, samplesData)

}
\author{
Samuel Wieczorek
}
