% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{createMSnset}
\alias{createMSnset}
\title{Creates an object of class \code{\link{MSnSet}} from text file}
\usage{
createMSnset(file, metadata = NULL, indExpData, indFData, indiceID = NULL,
  logData = FALSE, replaceZeros = FALSE, pep_prot_data = NULL)
}
\arguments{
\item{file}{The name of a tab-separated file that contains the data.}

\item{metadata}{A dataframe describing the samples (in lines).}

\item{indExpData}{A vector of string where each element is the name
of a column in designTable that have to be integrated in
the \code{fData()} table
of the \code{\link{MSnSet}} object.}

\item{indFData}{The name of column in \code{file} that will be the name of
rows for the \code{exprs()} and \code{fData()} tables}

\item{indiceID}{The indice of the column containing the ID of entities 
(peptides or proteins)}

\item{logData}{A boolean value to indicate if the data have to be
log-transformed (Default is FALSE)}

\item{replaceZeros}{A boolean value to indicate if the 0 and NaN values of
intensity have to be replaced by NA (Default is FALSE)}

\item{pep_prot_data}{A string that indicates whether the dataset is about 
peptides or proteins.}
}
\value{
An instance of class \code{\link{MSnSet}}.
}
\description{
Builds an object of class \code{\link{MSnSet}} from a 
single tabulated-like file for quantitative and meta-data and a dataframe 
for the samples description. It differs from
the original \code{MSnSet} builder which requires three separated files 
tabulated-like quantitative proteomic data into a \code{MSnSet} object,
including metadata.
}
\examples{
library(DAPARdata)
exprsFile <- system.file("extdata", "UPSprot25.txt", package="DAPARdata")
metadataFile <- system.file("extdata", "samples.txt", package="DAPARdata")
metadata = read.table(metadataFile, header=TRUE, sep="\\t", as.is=TRUE)
indExpData <- c(49:54)
indFData <- c(1:48, 55:86)
indiceID <- 78
createMSnset(exprsFile, metadata,indExpData,  indFData, indiceID,pep_prot_data = "protein")
}
\author{
Florence Combes, Samuel Wieczorek
}

