% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{CVDistD}
\alias{CVDistD}
\title{Distribution of CV of entities}
\usage{
CVDistD(qData, labels = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{labels}{A vector of the conditions (labels) (one label per sample).}
}
\value{
A density plot
}
\description{
Builds a densityplot of the CV of entities in the exprs() table
of a object. The CV is calculated for each condition (Label) present
in the dataset (see the slot \code{'Label'} in the \code{pData()} table)
}
\examples{
data(UPSpep25)
labels <- Biobase::pData(UPSpep25)[,"Label"]
CVDistD(UPSpep25)
}
\seealso{
\code{\link{densityPlotD}}.
}
\author{
Florence Combes, Samuel Wieczorek
}
