% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dapar_hc_chart}
\alias{dapar_hc_chart}
\title{Customised resetZoomButton of highcharts plots}
\usage{
dapar_hc_chart(hc, chartType, zoomType = "None", width = 0, height = 0)
}
\arguments{
\item{hc}{A highcharter object}

\item{chartType}{The type of the plot}

\item{zoomType}{The type of the zoom (one of "x", "y", "xy", "None")}

\item{width}{xxx}

\item{height}{xxx}
}
\value{
A highchart plot
}
\description{
Customised resetZoomButton of highcharts plots
}
\examples{
library("highcharter")
hc <- highchart()
hc <- dapar_hc_chart(hc, chartType = "line", zoomType = "x")
hc_add_series(hc, data = c(29, 71, 40))

}
\author{
Samuel Wieczorek
}
