% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dapar_hc_ExportMenu}
\alias{dapar_hc_ExportMenu}
\title{Selects lines of dataframes in a MSnset file}
\usage{
dapar_hc_ExportMenu(hc, filename)
}
\arguments{
\item{hc}{A highcharter object}

\item{filename}{The filename under which the plot has to be saved}

\item{obj}{A MSnset object}

\item{lineIndices}{The indices of lines to be extracted}
}
\value{
A MSnset object

A contextual menu for highcharts plots
}
\description{
This function overloads the brackets to select lines of dataframes in a MSnset file. It takes 
into account the slots experimentData / other / OriginOfValues
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
res <- tabOperator(Exp1_R25_pept, c(1:10))
library("highcharter")
hc <- highchart() 
hc_chart(hc,type = "line") 
hc_add_series(hc,data = c(29, 71, 40))
dapar_hc_ExportMenu(hc,filename='foo')

}
\author{
Samuel Wieczorek

Samuel Wieczorek
}
