% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{DAPARVersion,MSnSet-method}
\alias{DAPARVersion,MSnSet-method}
\title{Get the version of DAPAR used.}
\usage{
\S4method{DAPARVersion}{MSnSet}(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
xxxx.
}
\description{
qkljsdh qsodhqs qsidsq.
}
\details{
Function to get the version of DAPAR used to create the current object
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
