% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{DAPARVersion<-,MSnSet-method}
\alias{DAPARVersion<-,MSnSet-method}
\title{Set the version of DAPAR used.}
\usage{
\S4method{DAPARVersion}{MSnSet}(obj) <- value
}
\arguments{
\item{obj}{xxxx}

\item{value}{xxx}
}
\value{
xxxx.
}
\description{
qsdh qiosdhqs qisdq.
}
\details{
Function to set the version of DAPAR used to create the current object
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
