% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{DAPARVersion}
\alias{DAPARVersion}
\title{qfsqfds}
\usage{
DAPARVersion(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
xxxx.
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
