% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{dataset,PipelineTemplate-method}
\alias{dataset,PipelineTemplate-method}
\title{Get a dataset slot value for the object}
\usage{
\S4method{dataset}{PipelineTemplate}(obj, name)
}
\arguments{
\item{obj}{xxxx}

\item{name}{xxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
