% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{deleteLinesFromIndices}
\alias{deleteLinesFromIndices}
\title{Delete the lines in the matrix of intensities and the metadata table
given their indice.}
\usage{
deleteLinesFromIndices(obj, deleteThat = NULL, processText = NULL)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}} containing
quantitative data.}

\item{deleteThat}{A vector of integers which are the indices of lines to 
delete.}

\item{processText}{A string to be included in the \code{\link{MSnSet}}
object for log.}
}
\value{
An instance of class \code{\link{MSnSet}} that have been filtered.
}
\description{
Delete the lines of \code{exprs()} table identified by their indice.
}
\examples{
data(UPSpep25)
mvFilter(UPSpep25, c(1:10))
}
\author{
Florence Combes, Samuel Wieczorek
}

