% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{densityPlotD}
\alias{densityPlotD}
\title{Builds a densityplot from an object of class \code{\link{MSnSet}}}
\usage{
densityPlotD(obj, highLightLabel = NULL, lab2Show = NULL)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{highLightLabel}{The name of the Label to highlight
in the density plot.}

\item{lab2Show}{A vector of labels to show in densityplot.}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples.
}
\examples{
data(UPSprotx2)
densityPlotD(UPSprotx2)
}
\author{
Florence Combes, Samuel Wieczorek
}
\seealso{
\code{\link{boxPlotD}}, \code{\link{varianceDistD}}
}

