% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{designMap,PipelineTemplate-method}
\alias{designMap,PipelineTemplate-method}
\title{Get the designMap slot value for the object}
\usage{
\S4method{designMap}{PipelineTemplate}(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
The value of the slot called designMap
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
