% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAna}
\alias{diffAna}
\title{This function performs a differential analysis on an MSnSet
object (adapted from \code{\link{limma}})}
\usage{
diffAna(obj, design)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{design}{The design matrix as described in the limma package
documentation}
}
\value{
A dataframe with the p-value and log(Fold Change) associated
to each element (peptide/protein)
}
\description{
Performs a differential analysis on an \code{\link{MSnSet}} object,
based on \code{\link{limma}} functions.
}
\examples{
data(UPSprotx2)
design <- cbind(cond1=1, cond2 = rep(0,nrow(pData(UPSprotx2))))
rownames(design) <- rownames(pData(UPSprotx2))
indices <- getIndicesConditions(UPSprotx2, "10fmol", "5fmol")
design[indices$iCond2,2] <- 1
diffAna(UPSprotx2, design)
}
\author{
Florence Combes, Samuel Wieczorek
}

