% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaComputeFDR}
\alias{diffAnaComputeFDR}
\title{Computes the FDR corresponding to the p-values of the
differential analysis using}
\usage{
diffAnaComputeFDR(adj.pvals)
}
\arguments{
\item{adj.pvals}{xxxx}
}
\value{
The computed FDR value (floating number)
}
\description{
This function is a wrapper to the function adjust.p from the `cp4p` package.
 It returns the FDR corresponding to the p-values of the differential 
analysis. The FDR is computed with the function \code{p.adjust}\{stats\}.
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
