% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaSave}
\alias{diffAnaSave}
\title{Returns a \code{\link{MSnSet}} object with the results of
the differential analysis performed with \code{\link{limma}} package.}
\usage{
diffAnaSave(obj, data, method = "limma", condition1, condition2,
  threshold_pVal = 0, threshold_logFC = 0, fdr = 0)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{data}{The result of the differential analysis processed
by \code{\link{diffAna}}}

\item{method}{The method used for differential analysis.
Available choices are : "limma", "Welch"}

\item{condition1}{A vector containing the names (some values of the slot
"Label" of \code{pData()} of the first condition.}

\item{condition2}{A vector containing the names (some values of the slot
"Label" of \code{pData()} of the second condition.}

\item{threshold_pVal}{A float that indicates the threshold on p-value
choosen to discriminate differential proteins.}

\item{threshold_logFC}{A float that indicates the threshold on
log(Fold Change) to discriminatedifferential proteins.}

\item{fdr}{The FDR based on the values of threshold_pVal and
threshold_logFC}
}
\value{
A MSnSet
}
\description{
This method returns a \code{\link{MSnSet}} object with the results
of differential analysis.
}
\examples{
data(UPSprotx2)
condition1 <- '10fmol'
condition2 <- '5fmol'
limma <- diffAnaLimma(UPSprotx2, condition1, condition2)
obj <- diffAnaSave(UPSprotx2, limma, "limma", condition1, condition2)
}
\author{
Alexia Dorffer
}

