% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaSaveRAW_Data}
\alias{diffAnaSaveRAW_Data}
\title{Returns a \code{MSnSet} object with the RAW results of
the differential analysis}
\usage{
diffAnaSaveRAW_Data(obj, data, method = "limma")
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{data}{The result of the differential analysis processed 
by \code{\link{wrapper.limmaCompleteTest}}}

\item{method}{The method used for differential analysis.}
}
\value{
A MSnSet
}
\description{
This method returns a class \code{MSnSet} object with the RAW results (logFC and pValue for all
comparisons) of a differential analysis processed by \code{\link{wrapper.limmaCompleteTest}}.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
limma <- wrapper.limmaCompleteTest(Exp1_R25_pept[1:1000], 1)
obj <- diffAnaSaveRAW_Data(Exp1_R25_pept[1:1000], limma)
}
\author{
Samuel Wieczorek
}
