% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{display.CC}
\alias{display.CC}
\title{Display a CC}
\usage{
display.CC(The.CC, X, layout = layout_nicely)
}
\arguments{
\item{The.CC}{A cc (a list)}

\item{X}{xxxxx}

\item{layout}{xxxxx}
}
\value{
A plot
}
\description{
Display a CC
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept) 
X <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], "Protein.group.IDs", TRUE)
ll <- get.pep.prot.cc(X)
display.CC(ll[[1]], X)
}
\author{
Thomas Burger, Samuel Wieczorek
}
