% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{.checkIfAnalysisExists}
\alias{.checkIfAnalysisExists}
\title{Check if xxx}
\usage{
.checkIfAnalysisExists(object, name)
}
\arguments{
\item{object}{xxxx}

\item{name}{xxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
