% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{.ConfigureDataset}
\alias{.ConfigureDataset}
\title{Function to do some modifications of a MSnSet class dataset to be compatible with Prostar}
\usage{
.ConfigureDataset(x)
}
\description{
Function to do some modifications of a MSnSet class dataset to be compatible with Prostar
}
