% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{.validPipelineTemplate}
\alias{.validPipelineTemplate}
\title{Check if xxx}
\usage{
.validPipelineTemplate(object)
}
\arguments{
\item{object}{xxxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
