% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioAnalysis.R
\name{enrich_GO}
\alias{enrich_GO}
\title{Calculates the GO xxxxxxxxx}
\usage{
enrich_GO(data, idFrom, idTo, orgdb, ont, readable = TRUE, pval, universe)
}
\arguments{
\item{data}{A vector of ID (genes or proteins !!DIRE LESQUELS!!)}

\item{idFrom}{xxxxx}

\item{idTo}{xxxxx}

\item{orgdb}{xxxxx}

\item{ont}{One of "MF", "BP", and "CC" subontologies}

\item{readable}{xxxxx}

\item{pval}{The qvalue cutoff (same parameter as in the function \code{enrichGO} of the package \code{\link{clusterProfiler}})}

\item{universe}{xxxxxxxxx}
}
\value{
A groupGOResult instance.
}
\description{
This function is a wrappper to the function enrichGO from the
package \code{\link{clusterProfiler}}. Given a vector of genes/proteins, it returns the 
GO xxxxxxx.
}
\author{
xxxxx
}
