% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExtendPalette}
\alias{ExtendPalette}
\title{Extends a base-palette of the package RColorBrewer to n colors.}
\usage{
ExtendPalette(n = 0, base = "Set1")
}
\arguments{
\item{n}{The number of desired colors in the palette}

\item{base}{The name of the palette of the package RColorBrewer from which the extended palette is built.
Default value is 'Set1'.}
}
\value{
A vector composed of n color code.
}
\description{
The colors in the returned palette are always in the same order
}
\examples{
ExtendPalette(12)

}
\author{
Samuel Wieczorek
}
