% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{finalizeAggregation}
\alias{finalizeAggregation}
\title{xxxx}
\usage{
finalizeAggregation(obj.pep, qData, X, finalX, lib.loc = NULL)
}
\arguments{
\item{obj.pep}{xxxxx}

\item{qData}{xxxx}

\item{X}{xxxxx}

\item{finalX}{xxxxx}

\item{lib.loc}{xxxx}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\author{
Samuel Wieczorek
}
