% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{findMECBlock}
\alias{findMECBlock}
\title{Finds the LAPALA into a \code{MSnSet} object}
\usage{
findMECBlock(obj)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}
}
\value{
A data.frame that contains the indexes of LAPALA
}
\description{
Finds the LAPALA into a \code{MSnSet} object
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
obj <- Exp1_R25_pept[seq_len(100)]
lapala <- findMECBlock(obj)

}
\author{
Samuel Wieczorek
}
