% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{get.pep.prot.cc}
\alias{get.pep.prot.cc}
\title{Build the list of connex composant of the adjacency matrix}
\usage{
get.pep.prot.cc(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of CC
}
\description{
Build the list of connex composant of the adjacency matrix
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
obj <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj, "Protein_group_IDs", FALSE)
ll <- get.pep.prot.cc(X)

}
\author{
Thomas Burger, Samuel Wieczorek
}
