% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetCC}
\alias{GetCC}
\title{Returns the contains of the slot processing  of an object of 
class \code{MSnSet}}
\usage{
GetCC(obj)
}
\arguments{
\item{obj}{An object (peptides) of class \code{MSnSet}.}
}
\value{
A list of connected components
}
\description{
Returns the contains of the slot processing  of an object of 
class \code{MSnSet}
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
Xshared <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], "Protein_group_IDs", FALSE)
Xunique <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], "Protein_group_IDs", TRUE)
ll.X <- list(matWithSharedPeptides = Xshared, matWithUniquePeptides = Xunique)
Exp1_R25_pept <- SetMatAdj(Exp1_R25_pept, ll.X)
ll1 <- get.pep.prot.cc(GetMatAdj(Exp1_R25_pept)$matWithSharedPeptides)
ll2 <- DAPAR::get.pep.prot.cc(GetMatAdj(Exp1_R25_pept)$matWithUniquePeptides)
cc <- list(allPep = ll1, onlyUniquePep = ll2)
Exp1_R25_pept <- SetCC(Exp1_R25_pept, cc)
ll.cc <- GetCC(Exp1_R25_pept)

}
\author{
Samuel Wieczorek
}
