% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{GetColorsForConditions}
\alias{GetColorsForConditions}
\title{Builds a complete color palette for the conditions given in argument}
\usage{
GetColorsForConditions(conds, palette = NULL)
}
\arguments{
\item{conds}{The extended vector of samples conditions}

\item{palette}{A vector of HEX color code that form the basis palette from which to build
the complete color vector for the conditions.}
}
\value{
A vector composed of HEX color code for the conditions
}
\description{
xxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
conditions <- Biobase::pData(Exp1_R25_pept)$Condition
GetColorsForConditions(conditions, ExtendPalette(2))

}
\author{
Samuel Wieczorek
}
