% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{getDesignLevel}
\alias{getDesignLevel}
\title{xxx}
\usage{
getDesignLevel(sTab)
}
\arguments{
\item{sTab}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
sTab <- Biobase::pData(Exp1_R25_pept)
getDesignLevel(sTab)

}
