% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{GetDetailedNbPeptidesUsed}
\alias{GetDetailedNbPeptidesUsed}
\title{Computes the detailed number of peptides used for aggregating each protein}
\usage{
GetDetailedNbPeptidesUsed(X, pepData)
}
\arguments{
\item{X}{An adjacency matrix}

\item{pepData}{A data.frame of quantitative data}
}
\value{
A data.frame
}
\description{
Method to compute the detailed number of quantified peptides used for aggregating each protein
}
\author{
Samuel Wieczorek
}
