% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{GetIndices_Filter_By_Proportion}
\alias{GetIndices_Filter_By_Proportion}
\title{Filter missing values by proportion}
\usage{
GetIndices_Filter_By_Proportion(obj, intensities_proportion, mode = "None")
}
\arguments{
\item{obj}{An object of class \code{MSnSet} containing quantitative data
and phenotype data.}

\item{intensities_proportion}{float between 0 and 1 corresponding to the proportion
of intensities to keep in the lines.}

\item{mode}{character string. Four possibilities corresponding to the
description above: "None", wholeMatrix", "allCond" and "atLeastOneCond".}
}
\value{
the object given as input but with the lines not respecting the
proportion of NA requested in less.
}
\description{
Remove lines in the data according to the proportion of missing
values. This proportion is calculated differently depending on whether we
want a certain proportion of missing values (NA) to remain on:
* the entire matrix, regardless of the conditions: the rows containing a
proportion of NA equal or below the threshold will be kept.
* all the conditions: the lines for which all the conditions have a NA
proportion equal to or less than the fixed proportion will be kept.
* at least one condition: the lines for which at least one condition is
equal to or less than the fixed proportion of NA will be kept.
}
\examples{
utils::data(Exp1_R25_prot, package='DAPARdata')
filtered <- filterByProportion(obj = Exp1_R25_prot, intensities_proportion = 0.8, mode = "atLeastOneCond")

}
\author{
Hélène Borges, Samuel Wieczorek
}
