% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{GetIndices_WholeLine}
\alias{GetIndices_WholeLine}
\title{Search lines which respects query on all their elements.}
\usage{
GetIndices_WholeLine(metacell.mask)
}
\arguments{
\item{metacell.mask}{xxx}
}
\value{
xxx
}
\description{
This function looks for the lines where each element respect the query.
}
\examples{
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[seq.int(from=20, to=30)]
level <- 'peptide'
pattern <- "missing POV"
metacell.mask <- match.metacell(metadata = GetMetacell(obj), 
pattern = pattern, level = level)
ind <- GetIndices_WholeLine(metacell.mask)

}
