% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getIndicesConditions}
\alias{getIndicesConditions}
\title{Gets the conditions indices}
\usage{
getIndicesConditions(labels, cond1, cond2)
}
\arguments{
\item{labels}{A vector of strings containing the column "Label" of the \code{pData()}.}

\item{cond1}{A vector of Labels (a slot in the \code{pData()} table) for
the condition 1.}

\item{cond2}{A vector of Labels (a slot in the \code{pData()} table) for
the condition 2.}
}
\value{
A list with two slots \code{iCond1} and \code{iCond2} containing
respectively the indices of samples in the \code{pData()} table of the
dataset.
}
\description{
Returns a list for the two conditions where each slot is a vector of indices
for the samples
}
\examples{
library(DAPARdata)
data(UPSpep25)
labels <- pData(UPSpep25)[,"Label"]
getIndicesConditions(labels, "25fmol", "10fmol")
}
\author{
Florence Combes, Samuel Wieczorek
}

