% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{getIndicesOfLinesToRemove}
\alias{getIndicesOfLinesToRemove}
\title{xxxxxxxxx}
\usage{
getIndicesOfLinesToRemove(obj, idLine2Delete = NULL, prefix = NULL)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{idLine2Delete}{The name of the column that correspond to the data to filter}

\item{prefix}{A character string that is the prefix to find in the data}
}
\value{
A vector of integers.
}
\description{
Returns the indice of the lines in the dataset xxxxxxxxxx
}
\examples{
data(UPSpepx2)
getIndicesOfLinesToRemove(UPSpepx2, "Contaminant", prefix="+")
}
\author{
Samuel Wieczorek
}

