% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{GetNbPeptidesUsed}
\alias{GetNbPeptidesUsed}
\title{xxxx}
\usage{
GetNbPeptidesUsed(X, qData)
}
\arguments{
\item{X}{xxxx}

\item{qData}{xxxxx}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\author{
Samuel Wieczorek
}
