% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{GetNbTags}
\alias{GetNbTags}
\title{Number of each metacell tags}
\usage{
GetNbTags(obj)
}
\arguments{
\item{obj}{A instance of the class 'MSnset'}
}
\value{
xxx
}
\description{
Number of each metacell tags
}
\examples{
NULL
}
