% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{getNumberOfEmptyLines}
\alias{getNumberOfEmptyLines}
\title{Returns the number of empty lines in the data}
\usage{
getNumberOfEmptyLines(obj)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}
}
\value{
An integer
}
\description{
Returns the number of empty lines in the quantitative data
(i.e. \code{exprs()} table).
}
\examples{
data(UPSprotx2)
getNumberOfEmptyLines(UPSprotx2)
}
\author{
Samuel Wieczorek
}

