% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPaletteForConditions}
\alias{getPaletteForConditions}
\title{Palette for plots in \code{DAPAR}}
\usage{
getPaletteForConditions(conds)
}
\arguments{
\item{conds}{A vector of conditions (strings).}
}
\value{
A palette designed for the data manipulated in \code{DAPAR}
}
\description{
Selects colors for the plots in \code{DAPAR} based on the different conditions in
the dataset. The palette is derived from
the brewer palette "Dark2" (see \code{\link{RColorBrewer}}).
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
getPaletteForConditions(conds)
}
\author{
Florence Combes, Samuel Wieczorek
}
