% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getPaletteForLabels}
\alias{getPaletteForLabels}
\title{Palette for plots in DAPAR}
\usage{
getPaletteForLabels(obj)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}
}
\value{
A palette designed for the data manipulated in DAPAR
}
\description{
Selects colors for the plots in DAPAR based on the different conditions in
the dataset. The palette is derived from
the brewer palette "Dark2" (see \code{\link{RColorBrewer}}).
}
\examples{
data(UPSprotx2)
getPaletteForLabels(UPSprotx2)
}
\author{
Florence Combes, Samuel Wieczorek
}

