% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPaletteForReplicates}
\alias{getPaletteForReplicates}
\title{Palette for plot the replicates in DAPAR}
\usage{
getPaletteForReplicates(nColors)
}
\arguments{
\item{nColors}{The desired number of colors}
}
\value{
A palette designed for the data manipulated in DAPAR
}
\description{
Selects colors for the plots in DAPAR based on the replicates in
the dataset. The palette is derived from
the brewer palette "Dark2" (see \code{\link{RColorBrewer}}).
}
\examples{
data(UPSprotx2)
n <- nrow(pData(UPSprotx2))
getPaletteForLabels(5)
}
\author{
Samuel Wieczorek
}

