% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPaletteForReplicates_HC}
\alias{getPaletteForReplicates_HC}
\title{Palette for highcharter plot the replicates in \code{\link{DAPAR}}}
\usage{
getPaletteForReplicates_HC(nColors)
}
\arguments{
\item{nColors}{The desired number of colors}
}
\value{
A palette designed for the data manipulated in \code{\link{DAPAR}}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
n <- nrow(Biobase::pData(Exp1_R25_pept))
getPaletteForReplicates_HC(n)
}
\author{
Samuel Wieczorek
}
