% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{getPourcentageOfMV}
\alias{getPourcentageOfMV}
\title{Percentage of missing values}
\usage{
getPourcentageOfMV(obj)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}
}
\value{
A floating number
}
\description{
Returns the percentage of missing values in the quantitative
data (\code{exprs()} table of the dataset).
}
\examples{
library(DAPARdata)
data(UPSprotx2)
getPourcentageOfMV(UPSprotx2)
}
\author{
Florence Combes, Samuel Wieczorek
}

