% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getProcessingInfo}
\alias{getProcessingInfo}
\title{Returns the contains of the slot processing  of an object of 
class MSnSet}
\usage{
getProcessingInfo(obj)
}
\arguments{
\item{obj}{An object (peptides) of class \code{\link{MSnbase}}.}
}
\value{
The slot processing of obj@processingData
}
\description{
Returns the contains of the slot processing  of an object of class MSnSet.
}
\examples{
data(UPSpep25)
getProcessingInfo(UPSpep25)
}
\author{
Samuel Wieczorek
}

