% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{getProteinsStats}
\alias{getProteinsStats}
\title{computes the number of proteins that are only defined by 
specific peptides, shared peptides or a mixture of two.}
\usage{
getProteinsStats(matUnique, matShared)
}
\arguments{
\item{matUnique}{The adjacency matrix with only specific peptides.}

\item{matShared}{The adjacency matrix with both specific and shared peptides.}
}
\value{
A list
}
\description{
This function computes the number of proteins that are only defined by 
specific peptides, shared peptides or a mixture of two.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
protID <- "Protein.group.IDs"
MShared <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], protID, FALSE)
MUnique <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], protID, TRUE)
getProteinsStats(MUnique,MShared)
}
\author{
Samuel Wieczorek
}
