% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{getQuantile4Imp}
\alias{getQuantile4Imp}
\title{xxxxxxxxx}
\usage{
getQuantile4Imp(qData, qval = 0.025, factor = 1)
}
\arguments{
\item{qData}{xxxxxxxxx}

\item{qval}{xxxxxxxxx}

\item{factor}{xxxxxxxxx.}
}
\value{
xxxxxxxxx
}
\description{
This method is xxxxxxxx
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
getQuantile4Imp(qData) 
}
\author{
Thomas Burger
}
