% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{GetSoftAvailables}
\alias{GetSoftAvailables}
\title{The set of softwares available}
\usage{
GetSoftAvailables()
}
\value{
xxx
}
\description{
The set of softwares available
}
\examples{
GetSoftAvailables()

}
