% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForAggregation}
\alias{getTextForAggregation}
\title{Build the text information for the Aggregation process}
\usage{
getTextForAggregation(l.params)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
Builds the text information for the Aggregation process
}
\examples{
getTextForAggregation(list(POV_algorithm="slsa",MEC_algorithm="fixedValue", MEC_fixedValue = 0))
}
\author{
Samuel Wieczorek
}
