% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForFiltering}
\alias{getTextForFiltering}
\title{Build the text information for the filtering process}
\usage{
getTextForFiltering(l.params)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
Build the text information for the filtering process
}
\examples{
getTextForFiltering(list(mvFilterType="WholeMatrix",mvThNA=3, val_vs_percent='Value'))

}
\author{
Samuel Wieczorek
}
