% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForImputation}
\alias{getTextForImputation}
\title{Build the text information for the Imputation process}
\usage{
getTextForImputation(l.params, level)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}

\item{level}{The type of data (peptide or protein)}
}
\value{
A string
}
\description{
Build the text information for the Imputation process
}
\examples{
getTextForImputation(list(POV_algorithm="slsa",MEC_algorithm="fixedValue", MEC_fixedValue = 0), level="protein")
}
\author{
Samuel Wieczorek
}
