% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForpeptideImputation}
\alias{getTextForpeptideImputation}
\title{Build the text information for the peptide Imputation process}
\usage{
getTextForpeptideImputation(l.params)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
* pepLevel_algorithm,
* pepLevel_basicAlgorithm,
* pepLevel_detQuantile,
* pepLevel_detQuant_factor,
* pepLevel_imp4p_nbiter,
* pepLevel_imp4p_withLapala,
* pepLevel_imp4p_qmin,
* pepLevel_imp4pLAPALA_distrib
}
\examples{
params <- list()
getTextForpeptideImputation(params)

}
\author{
Samuel Wieczorek
}
