% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForproteinImputation}
\alias{getTextForproteinImputation}
\title{Build the text information for the protein Imputation process}
\usage{
getTextForproteinImputation(l.params)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
* POV_algorithm,
* POV_detQuant_quantile,
* POV_detQuant_factor,
* POV_KNN_n,
* MEC_algorithm,
* MEC_detQuant_quantile,
* MEC_detQuant_factor,
* MEC_fixedValue
}
\examples{
params <- list()
getTextForproteinImputation(params)

}
\author{
Samuel Wieczorek
}
