% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetTypeofData}
\alias{GetTypeofData}
\title{xxxx}
\usage{
GetTypeofData(obj)
}
\arguments{
\item{obj}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
GetTypeofData(Exp1_R25_pept)

}
