% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioAnalysis.R
\name{GOAnalysisSave}
\alias{GOAnalysisSave}
\title{xxxxxxxxxxxxxx}
\usage{
GOAnalysisSave(obj, ggo_res, ego_res, organism, ontology, level, PAdjustMethod,
  pvalueCutoff, typeUniverse)
}
\arguments{
\item{dat}{xxxxxxxxxxxxxx}
}
\value{
xxxxxxxxxxxxxx
}
\description{
xxxxxxxxxxxxxx
}
\author{
Samuel Wieczorek
}
