% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ttest.R
\name{groupttestsw}
\alias{groupttestsw}
\title{xxxxxx}
\usage{
groupttestsw(MatAdj, cond1, cond2)
}
\arguments{
\item{qData}{A matrix of quantitative data, without any missing values.}

\item{X}{A matrix of quantitative data, without any missing values.}
}
\value{
xxxxx
}
\description{
This function is xxxxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept
protID <- "Protein_group_IDs"
keepThat <-  mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
X.spec <- BuildAdjacencyMatrix(obj, protID,  unique = TRUE)
qData <- Biobase::exprs(obj)
gttest <- groupttest(X.spec, qData)
}
\author{
Thomas Burger, Samuel Wieczorek
}
