% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{hc_FC_DensityPlot}
\alias{hc_FC_DensityPlot}
\title{Density plots of FC values}
\usage{
hc_FC_DensityPlot(df_FC, threshold_LogFC = 0)
}
\arguments{
\item{df_FC}{A dataframe that contains the FC values}

\item{threshold_LogFC}{The threshold on log(Fold Change) to
distinguish between differential and non-differential data}
}
\value{
A highcharts density plot
}
\description{
This function show the density plots of Fold Change (the same as calculated by limma) for a list 
of the comparisons of conditions in a differnetial analysis.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[1:1000]
lapala <- findMECBlock(obj)
obj <- wrapper.impute.detQuant(obj)
obj <- reIntroduceMEC(obj, lapala)
obj <- wrapper.impute.detQuant(obj)
limma <- wrapper.limmaCompleteTest(obj, 1)
hc_FC_DensityPlot(limma$FC)
}
\author{
Samuel Wieczorek
}
