% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{heatmapD}
\alias{heatmapD}
\title{This function is a wrapper to \code{\link{heatmap.2}} that displays
quantitative data in the \code{exprs()} table of an object of
class \code{\link{MSnSet}}}
\usage{
heatmapD(obj, distance = "euclidean", cluster = "average")
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{distance}{The distance used by the clustering algorithm to compute
the dendrogram. See \code{help(heatmap.2)}}

\item{cluster}{the clustering algorithm used to build the dendrogram.
See \code{help(heatmap.2)}}
}
\value{
A heatmap
}
\description{
Heatmap of the quantitative proteomic data of a \code{\link{MSnSet}} object
}
\examples{
data(testWithoutNA)
heatmapD(testWithoutNA)
}
\author{
Florence Combes, Samuel Wieczorek
}

