% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{histPValue_HC}
\alias{histPValue_HC}
\title{Plots a histogram ov p-values}
\usage{
histPValue_HC(pval_ll, bins = 80, pi0 = 1)
}
\arguments{
\item{pval_ll}{xxx}

\item{bins}{xxx}

\item{pi0}{xxx}
}
\value{
A plot
}
\description{
This function plots a histogram ov p-values
}
\examples{
utils::data(Exp1_R25_prot, package='DAPARdata')
obj <- Exp1_R25_prot[1:1000]
level <- obj@experimentData@other$typeOfData
metacell.mask <- match.metacell(GetMetacell(obj), 'missing', level)
indices <- GetIndices_WholeMatrix(metacell.mask, op='>=', th=1)
obj <- MetaCellFiltering(obj, indices, cmd='delete')
qData <- exprs(obj$new)
sTab <- pData(obj$new)
allComp <- limmaCompleteTest(qData,sTab)
histPValue_HC(allComp$P_Value[1])

}
\author{
Samuel Wieczorek
}
