% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{impute.detQuant}
\alias{impute.detQuant}
\title{Deterministic imputation}
\usage{
impute.detQuant(qData, values, na.type = NULL)
}
\arguments{
\item{qData}{An expression set containing quantitative or missing values}

\item{values}{A vector with as many elements as the number of colums of qData}

\item{na.type}{xxx}
}
\value{
An imputed dataset
}
\description{
This method replaces each missing value by a given value
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
qData <- Biobase::exprs(Exp1_R25_pept)
values <- getQuantile4Imp(qData)$shiftedImpVal
impute.detQuant(qData, values, na.type = 'MEC') 

}
\author{
Thomas Burger, Samuel Wieczorek
}
