% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.topn}
\alias{inner.aggregate.topn}
\title{xxxx}
\usage{
inner.aggregate.topn(pepData, X, method = "Mean", n = 10)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}

\item{method}{xxxxx}

\item{n}{xxxxx}
}
\value{
xxxxx
}
\description{
xxxx
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
obj <- Exp1_R25_pept[seq_len(10)]
protID <- "Protein_group_IDs"
X <- BuildAdjacencyMatrix(obj, protID, FALSE)
inner.aggregate.topn(Biobase::exprs(obj), X)

}
\author{
Samuel Wieczorek
}
