% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.mean}
\alias{inner.mean}
\title{xxxx}
\usage{
inner.mean(pepData, X)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
xxxxx
}
\description{
xxxx
}
\examples{
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[seq_len(10)]
protID <- "Protein_group_IDs"
X <- BuildAdjacencyMatrix(obj, protID, FALSE)
inner.mean(Biobase::exprs(obj), X)

}
\author{
Samuel Wieczorek
}
