% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.sum}
\alias{inner.sum}
\title{xxxx}
\usage{
inner.sum(pepData, X)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
A matrix
}
\description{
xxxx
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
obj <- Exp1_R25_pept[seq_len(10)]
protID <- "Protein_group_IDs"
X <- BuildAdjacencyMatrix(obj, protID, FALSE)
inner.sum(Biobase::exprs(obj), X)
}
\author{
Samuel Wieczorek
}
