% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.byMSMS}
\alias{is.byMSMS}
\title{Similar to the function \code{is.na} but focused on the equality with the identification of the peptides ('By MS/MS')}
\usage{
is.byMSMS(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A boolean dataframe
}
\description{
Similar to the function \code{is.na} but focused on the equality with the identification of the peptides ('By MS/MS')
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept
data <- Biobase::fData(obj)[,obj@experimentData@other$OriginOfValues]
is.byMSMS(data)

}
\author{
Samuel Wieczorek, Enora Fremy
}
