% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.subset}
\alias{is.subset}
\title{xxx}
\usage{
is.subset(set1, set2)
}
\arguments{
\item{set1}{xxx}

\item{set2}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
is.subset('a', letters)
is.subset(c('a', 'c', 't'), letters)
is.subset(c('a', 3, 't'), letters)
is.subset(3, letters)

}
