% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{keyId,MSnSet-method}
\alias{keyId,MSnSet-method}
\title{Get the keyId value for the object (peptide)}
\usage{
\S4method{keyId}{MSnSet}(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
Th value of parentProtId which is a column of the dataset which contains xxxx.
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
