% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{make.design.2}
\alias{make.design.2}
\title{Builds the design matrix for designs of level 2}
\usage{
make.design.2(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the `pData()` function 
of package `MSnbase`.}
}
\value{
A design matrix
}
\description{
Builds the design matrix for designs of level 2
}
\examples{
data(Exp1_R25_pept, package='DAPARdata')
make.design.2(Biobase::pData(Exp1_R25_pept))


}
\author{
Thomas Burger, Quentin Giai-Gianetto, Samuel Wieczorek
}
