% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{MeanPeptidesSpeedUp}
\alias{MeanPeptidesSpeedUp}
\title{Compute the intensity of proteins as the mean of the intensities
of their peptides.}
\usage{
MeanPeptidesSpeedUp(matAdj, expr)
}
\arguments{
\item{matAdj}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{expr}{A matrix of intensities of peptides}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins as the mean of the 
intensities of their peptides.
}
\examples{
data(UPSpep25)
protID <- "Protein.group.IDs"
matAdj <- BuildSparseAdjacencyMatrix(UPSpep25, protID, FALSE)
MeanPeptidesSpeedUp(matAdj, Biobase::exprs(UPSpep25))
}
\author{
Alexia Dorffer
}

