% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{Metacell_generic}
\alias{Metacell_generic}
\title{Sets the metacell dataframe}
\usage{
Metacell_generic(qdata, conds, level)
}
\arguments{
\item{qdata}{An object of class \code{MSnSet}}

\item{conds}{xxx}

\item{level}{xxx}
}
\value{
xxxxx
}
\description{
In the quantitative columns, a missing value is identified by no value rather
than a value equal to 0. 
Conversion rules
Quanti			Tag		
NA or 0		NA
}
\examples{
file <- system.file("extdata", "Exp1_R25_pept.txt", package="DAPARdata")
data <- read.table(file, header=TRUE, sep="\t",stringsAsFactors = FALSE)
metadataFile <- system.file("extdata", "samples_Exp1_R25.txt", package="DAPARdata")
metadata <- read.table(metadataFile, header=TRUE, sep="\t", as.is=TRUE, stringsAsFactors = FALSE)
conds <- metadata$Condition
qdata <- data[,56:61]
df <- data[ , 43:48]
df <- Metacell_generic(qdata, conds, level='peptide')

}
\author{
Samuel Wieczorek
}
