% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{mvFilterGetIndices_Marianne}
\alias{mvFilterGetIndices_Marianne}
\title{Filter lines in the matrix of intensities w.r.t. some criteria}
\usage{
mvFilterGetIndices_Marianne(
  obj,
  percent = FALSE,
  condition = "WholeMatrix",
  threshold = NULL
)
}
\arguments{
\item{obj}{An object of class \code{MSnSet} containing
quantitative data.}

\item{percent}{TRUE or FALSE. Default is FALSE.}

\item{condition}{Method used to choose the lines to delete.
Values are : "None", "WholeMatrix", "AllCond", "AtLeastOneCond"}

\item{threshold}{An integer value of the threshold if percent is FALSE. 
Otherwise, a floating number between 0 and 1.}
}
\value{
An vector of indices that correspond to the lines to keep.
}
\description{
Returns the indices of the lines of \code{exprs()} table to delete w.r.t. 
the conditions on the number of missing values.
The user chooses the minimum amount of peptides identified by MS/MS that is 
acceptable and the filter delete lines that do not respect this condition.
The condition may be on the whole line or condition by condition.
}
\details{
The different methods are :
"WholeMatrix": given a threshold \code{th}, only the lines that contain
at least \code{th} 'by MS/MS' are kept.
"AllCond": given a threshold \code{th}, only the lines which contain
at least \code{th} 'by MS/MS' for each of the conditions are kept.
"AtLeastOneCond": given a threshold \code{th}, only the lines that contain
at least \code{th} 'by MS/MS', and for at least one condition, are kept.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj<-Exp1_R25_pept[1:100]
ind <- mvFilterGetIndices_Marianne(obj, condition = "WholeMatrix", 
threshold=1)
ind <- mvFilterGetIndices_Marianne(obj, condition = "EmptyLines")
ind <-mvFilterGetIndices_Marianne(obj, condition = "WholeMatrix", 
threshold=2)
ind <-mvFilterGetIndices_Marianne(obj, condition = "AllCond",  threshold=1)
ind <-mvFilterGetIndices_Marianne(obj,condition = "AtLeastOneCond",
threshold=1)

}
\author{
Enora Fremy, Samuel Wieczorek
}
