% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{mvHisto}
\alias{mvHisto}
\title{Histogram of missing values}
\usage{
mvHisto(qData, samplesData, labels, indLegend = "auto", showValues = FALSE)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{samplesData}{A dataframe where lines correspond to samples and columns to the meta-data for those samples.}

\item{labels}{A vector of the conditions (labels) (one label per sample).}

\item{indLegend}{The indices of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A histogram
}
\description{
This method plots a histogram of missing values.
}
\examples{
library(DAPARdata)
data(UPSpep25)
qData <- exprs(UPSpep25)
samplesData <- pData(UPSpep25)
labels <- pData(UPSpep25)[,"Label"]
mvHisto(qData, samplesData, labels, indLegend="auto", showValues=TRUE)
}
\author{
Florence Combes, Samuel Wieczorek
}

