% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{mvHisto_HC}
\alias{mvHisto_HC}
\title{Histogram of missing values}
\usage{
mvHisto_HC(qData, samplesData, labels, indLegend = "auto",
  showValues = FALSE)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{samplesData}{A dataframe where lines correspond to samples and 
columns to the meta-data for those samples.}

\item{labels}{A vector of the conditions (labels) (one label per sample).}

\item{indLegend}{The indices of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A histogram
}
\description{
This method plots a histogram of missing values. Same as the function \code{mvHisto}
but uses the package \code{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
samplesData <- Biobase::pData(Exp1_R25_pept)
labels <- Biobase::pData(Exp1_R25_pept)[,"Label"]
mvHisto_HC(qData, samplesData, labels, indLegend="auto", showValues=TRUE)
}
\author{
Florence Combes, Samuel Wieczorek
}
