% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{mvImputation}
\alias{mvImputation}
\title{Missing values imputation}
\usage{
mvImputation(obj, method)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{method}{The imputation method to be used.
Choices are QRILC, KNN, BPCA and MLE.}
}
\value{
The object \code{obj} which has been imputed
}
\description{
This method is a wrapper to the \code{imputeLCMD} package adapted to
objects of class \code{\link{MSnSet}}.
}
\examples{
data(UPSprotx2)
mvImputation(UPSprotx2, "QRILC")
}
\author{
Samuel Wieczorek
}

