% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{mvPerLinesHisto}
\alias{mvPerLinesHisto}
\title{Histogram of missing values per lines}
\usage{
mvPerLinesHisto(obj, indLegend = "auto", showValues = TRUE)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{indLegend}{The indice of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A histogram
}
\description{
This method plots a histogram which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\examples{
data(UPSprotx2)
mvPerLinesHisto(UPSprotx2)
}
\author{
Florence Combes, Samuel Wieczorek
}

