% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{mvPerLinesHistoPerCondition}
\alias{mvPerLinesHistoPerCondition}
\title{Bar plot of missing values per lines and per condition}
\usage{
mvPerLinesHistoPerCondition(qData, samplesData, indLegend = "auto",
  showValues = FALSE)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{samplesData}{A dataframe where lines correspond to samples and columns to the meta-data for those samples.}

\item{indLegend}{The indice of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A bar plot
}
\description{
This method plots a bar plot which represents the distribution of the 
number of missing values (NA) per lines (ie proteins) and per conditions.
}
\examples{
library(DAPARdata)
data(UPSprotx2)
qData <- exprs(UPSprotx2)
samplesData <- pData(UPSprotx2)
mvPerLinesHistoPerCondition(qData, samplesData)
}
\author{
Samuel Wieczorek
}

