% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{mvTypePlot}
\alias{mvTypePlot}
\title{Distribution of missing values with respect to intensity values}
\usage{
mvTypePlot(qData, labels, threshold = 0)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{labels}{A vector of the conditions (labels) (one label per sample).}

\item{threshold}{An integer for the intensity that delimits MNAR and 
MCAR missing values.}
}
\value{
A scatter plot
}
\description{
This method plots a scatter plot which represents the distribution of
missing values.
The colors correspond to the different conditions (slot Label in in the
dataset of class \code{\link{MSnSet}}).
The x-axis represent the mean of intensity for one condition and one
entity in the dataset (i. e. a protein) 
whereas the y-axis count the number of missing values for this entity
and the considered condition.
The data have been jittered for an easier vizualisation.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
labels <- Biobase::pData(Exp1_R25_pept)[,"Label"]
mvTypePlot(qData, labels, threshold=0)
}
\author{
Florence Combes, Samuel Wieczorek
}
