% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeD2}
\alias{normalizeD2}
\title{Normalisation}
\usage{
normalizeD2(qData, labels, method, type, scaling = FALSE, quantile = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{labels}{A vector of strings containing the column "Label" of 
the \code{pData()}.}

\item{method}{One of the following : Global Rescaling, 
Quantile Centering, Mean Centering.}

\item{type}{"Overall" or "Within conditions".}

\item{scaling}{A boolean}

\item{quantile}{A float}
}
\value{
A matrix normalized
}
\description{
Provides several methods to normalize data from
a matrix.
They are organized in four main families : Strong Rescaling, 
Median Centering, Mean Centering, Mean CenteringScaling.
For the first family, two sub-categories are available : the sum by columns
and the quantiles method.
For the three other families, two categories are available : 
"Overall" which means that the value for each protein 
(ie line in the expression data tab) is computed over all the samples ;
"within conditions" which means that the value for each protein 
(ie line in the matrix) is computed condition
by condition.
}
\examples{
data(UPSpep25)
qData <- Biobase::exprs(UPSpep25)
labels <- Biobase::pData(UPSpep25)[,"Label"]
normalizeD(qData, labels, "Median Centering", "within conditions")
}
\author{
Florence Combes, Samuel Wieczorek, Thomas Burger
}
