% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{OriginOfValues,MSnSet-method}
\alias{OriginOfValues,MSnSet-method}
\title{Get the OriginOfValues value for the object (peptide)}
\usage{
\S4method{OriginOfValues}{MSnSet}(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
The value of OriginOfValues which is a column of the dataset which contains xxxx.
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
