% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{OWAnova}
\alias{OWAnova}
\title{xxx}
\usage{
OWAnova(current_protein, conditions)
}
\arguments{
\item{current_protein}{xxx}

\item{conditions}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
NULL

}
\author{
Thomas Burger
}
