% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{PairwiseComps,PipelineTemplate-method}
\alias{PairwiseComps,PipelineTemplate-method}
\title{Get the PairwiseComps slot value for the object}
\usage{
\S4method{PairwiseComps}{PipelineTemplate}(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
The value of the slot called PairwiseComps
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
