% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{Params,MSnSet-method}
\alias{Params,MSnSet-method}
\title{Set the Params value for the object}
\usage{
\S4method{Params}{MSnSet}(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
The value of Params which is a column of the dataset which contains xxxx.
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
