% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{parentProtId<-,MSnSet-method}
\alias{parentProtId<-,MSnSet-method}
\title{Set the parentProtId value for the object (peptide)}
\usage{
\S4method{parentProtId}{MSnSet}(obj) <- value
}
\arguments{
\item{obj}{xxxx}

\item{value}{xxx}
}
\value{
xxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
