% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{parentProtId<-}
\alias{parentProtId<-}
\title{qfsqfds}
\usage{
parentProtId(obj) <- value
}
\arguments{
\item{obj}{xxxx}

\item{value}{xxx}
}
\value{
xxxx.
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
