% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepa.R
\name{pepa.test}
\alias{pepa.test}
\title{PEptide based Protein differential Abundance test}
\usage{
pepa.test(X, y, n1, n2)
}
\arguments{
\item{X}{Binary q x p design matrix for q peptides and p}

\item{y}{q x n matrix representing the log intensities of q peptides}

\item{n1}{number of samples under condition 1. It is assumed that the first n1 columns of y}

\item{n2}{number of samples under condition 2.}
}
\value{
A list of the following elements:
}
\description{
This function is PEptide based Protein differential Abundance test
}
\author{
Thomas Burger, Laurent Jacob
}
