% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{pepAgregateSpeedUp}
\alias{pepAgregateSpeedUp}
\title{Function agregate peptides to proteins}
\usage{
pepAgregateSpeedUp(obj.pep, protID, method = "sum overall", matAdj = NULL,
  n = NULL)
}
\arguments{
\item{obj.pep}{An object (peptides) of class \code{\link{MSnbase}}.}

\item{protID}{The name of proteins ID column}

\item{method}{The method used to aggregate the peptides into proteins.
Values are "sum", "mean" or "sum on top n" : do the sum / mean of intensity
on all peptides belonging to proteins. Default is "sum"}

\item{matAdj}{An adjacency matrix}

\item{n}{The number of peptides considered for the aggregation.}
}
\value{
An object of class \code{\link{MSnbase}} with proteins
}
\description{
Method to agregate with a method peptides to proteins on
a MSnSet object (peptides)
}
\examples{
data(UPSpep25)
protID <- "Protein.group.IDs"
mat <- BuildSparseAdjacencyMatrix(UPSpep25, protID, TRUE)
pepAgregateSpeedUp(UPSpep25, protID, "sum overall", mat)
}
\author{
Alexia Dorffer, Samuel Wieczorek
}

