% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-pipeline-class.R
\docType{class}
\name{PipelinePeptide-class}
\alias{PipelinePeptide-class}
\alias{.PipelinePeptide}
\title{PipelinePeptide class}
\value{
A \code{PipelinePeptide} object
}
\description{
xxxxx
}
\section{Slots}{

\describe{
\item{\code{matAdj}}{A character vector that is the name of the MS study}

\item{\code{CC}}{A character vector that indicates the type of data that are managed in this instance}
}}

\examples{
example("PipelineTemp")
library(DAPARdata)
data('Exp1_R25_pept')
obj <- Exp1_R25_pept
samples <- Biobase::pData(obj)
mae <- PipelinePeptide(analysis= 'test',pipelineType = 'peptide', dataType = 'peptide',
processes='original',experiments=list(original=obj), 
colData=samples)
}
