% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-pipeline-class.R
\name{PipelinePeptide}
\alias{PipelinePeptide}
\title{sdfdsfs}
\usage{
PipelinePeptide(...)
}
\arguments{
\item{...}{xxxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
