% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-pipeline-class.R
\docType{class}
\name{PipelineProtein-class}
\alias{PipelineProtein-class}
\alias{.PipelineProtein}
\title{PipelineProtein class}
\arguments{
\item{...}{Additional arguments for supporting functions. See details.}
}
\value{
A \code{PipelineProtein} object
}
\description{
xx x
}
\examples{
example("PipelineTemp")
library(DAPARdata)
data('Exp1_R25_prot')
samples <- Biobase::pData(Exp1_R25_prot)
mae <- PipelineProtein(analysis= 'test',pipelineType = 'protein', dataType = 'protein',
processes='original',experiments=list(original=Exp1_R25_prot), 
colData=samples)
}
