% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\docType{class}
\name{PipelineTemplate-class}
\alias{PipelineTemplate-class}
\alias{.PipelineTemplate}
\title{PipelineTemplate class}
\arguments{
\item{...}{Additional arguments for supporting functions. See details.}
}
\value{
A \code{PipelineTemplate} object
}
\description{
The \code{PipelineTemplate} class inherits from the \code{MultiAssayExperiment} and serves as a template for
instanciate pipeline classes such as proteinPipeline, proteinPipeline, etc. It can be used to manage results of
diverse assays on a collection of specimen. Currently,  the class can handle
assays that are organized instances of
\code{\linkS4class{MSnSet}},
\code{matrix}.
}
\section{Slots}{

\describe{
\item{\code{PairwiseComparisons}}{A \code{list} to store the result of hypothesis tests.}

\item{\code{indexNA}}{A xxxxxs}

\item{\code{dataType}}{xxxx.}

\item{\code{analysis}}{A character vector that is the name of the MS study}

\item{\code{pipelineType}}{A character vector that indicates the type of data that are managed in this instance}

\item{\code{processes}}{xxx}

\item{\code{version}}{xxx}
}}

\examples{
example("PipelineTemp")
library(DAPARdata)
data('Exp1_R25_prot')
obj <- Exp1_R25_prot
samples <- Biobase::pData(Exp1_R25_prot)
mae <- PipelineTemplate(analysis= 'test',pipelineType = 'protein', dataType = 'protein', 
processes='original',experiments=list(original=Exp1_R25_prot), 
colData=samples)
}
