% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{PipelineTemplate}
\alias{PipelineTemplate}
\title{xxxxx.}
\usage{
PipelineTemplate(
  PairwiseComparisons = list(),
  analysis = character(),
  dataType = character(),
  pipelineType = character(),
  processes = character(),
  version = character(),
  ...
)
}
\arguments{
\item{PairwiseComparisons}{xxxx}

\item{analysis}{A string that is the name of the object (ie the name of the MS study)}

\item{dataType}{A string that is the type of data}

\item{pipelineType}{the type of the pipeline used in Prostar}

\item{processes}{a character vector that indicates the list of the processes which composed the pipeline}

\item{version}{The version of Prostar used to create this object}

\item{...}{sdfkjdshfhdsfk}
}
\value{
An instance of class \code{PipelineTemplate}.
}
\description{
Function to create a new instance of the class \code{PipelineTemplate}
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
