% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{pipelineType,PipelineTemplate-method}
\alias{pipelineType,PipelineTemplate-method}
\title{Get the pipelineType slot value for the object}
\usage{
\S4method{pipelineType}{PipelineTemplate}(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
The value of the slot called pipelineType
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
