% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{processes<-,PipelineTemplate-method}
\alias{processes<-,PipelineTemplate-method}
\title{Set the processes slot value for the object}
\usage{
\S4method{processes}{PipelineTemplate}(obj) <- value
}
\arguments{
\item{obj}{xxxx}

\item{value}{xxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
