% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{proportionConRev}
\alias{proportionConRev}
\title{Barplot of proportion of contaminants and reverse}
\usage{
proportionConRev(obj, idContaminants = NULL, prefixContaminants = NULL,
idReverse = NULL, prefixReverse = NULL)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{idContaminants}{The name of a column of Contaminants}

\item{prefixContaminants}{The prefix to identify contaminants}

\item{idReverse}{The name of a column of Reverse}

\item{prefixReverse}{The prefix to identify Reverse}
}
\value{
A barplot
}
\description{
Plots a barplot of proportion of contaminants and reverse
}
\examples{
data(UPSpep25)
pref <- "+"
proportionConRev(UPSpep25, "Potential.contaminant", pref, "Reverse", pref)
}
\author{
Samuel Wieczorek
}

