% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{ProstarVersion,MSnSet-method}
\alias{ProstarVersion,MSnSet-method}
\title{Get the version of Prostar used.}
\usage{
\S4method{ProstarVersion}{MSnSet}(obj, withDimnames = TRUE)
}
\arguments{
\item{obj}{xxxx}

\item{value}{xxx}
}
\value{
xxxx.
}
\description{
sfgsdff sjkfhksqdsq .
}
\details{
Function to get the version of Prostar used to create the current object
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
