% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{ProstarVersion<-,MSnSet-method}
\alias{ProstarVersion<-,MSnSet-method}
\title{Set the version of Prostar used.}
\usage{
\S4method{ProstarVersion}{MSnSet}(obj) <- value
}
\arguments{
\item{obj}{xxxx}

\item{value}{xxx}
}
\value{
xxxx.
}
\description{
sfklsjhf qsjdhsqd.
}
\details{
Function to set the version of Prostar used to create the current object
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
