% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_for_MSnSet_class.R
\name{RawPValues<-,MSnSet-method}
\alias{RawPValues<-,MSnSet-method}
\title{Set the RawPValues value for the object}
\usage{
\S4method{RawPValues}{MSnSet}(obj) <- value
}
\arguments{
\item{obj}{xxxx}

\item{value}{xxx}
}
\value{
The value of RawPValues which is a column of the dataset which contains xxxx.
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
