% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{removeLines}
\alias{removeLines}
\title{xxxxxxxxx}
\usage{
removeLines(obj, idLine2Delete = NULL, prefix = NULL)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{idContaminants}{The name of the column that correspond to the boolean tag of contaminants}
}
\value{
An object of class \code{\link{MSnSet}}.
}
\description{
Removes lines in the dataset where the column Contaminants is TRUE
}
\examples{
data(UPSpepx2)
removeLines(UPSpepx2, "Contaminant")
removeLines(UPSpepx2, "Reverse")
}
\author{
Samuel Wieczorek
}

