% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{rmDatasetByIndice,PipelineTemplate-method}
\alias{rmDatasetByIndice,PipelineTemplate-method}
\title{Set the rmDatasetByIndice slot value for the object}
\usage{
\S4method{rmDatasetByIndice}{PipelineTemplate}(obj, ind)
}
\arguments{
\item{obj}{xxxx}

\item{ind}{xxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
