% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{search.metacell.tags}
\alias{search.metacell.tags}
\title{Search pattern in metacell vocabulary}
\usage{
search.metacell.tags(pattern, level)
}
\arguments{
\item{pattern}{The string to search.}

\item{level}{The available levels are : names()}
}
\description{
Gives all the tags of the metadata vocabulary containing the pattern 
(parent and all its children).
}
\examples{
search.metacell.tags('POV', 'peptide')
search.metacell.tags('MV_POV', 'peptide')
search.metacell.tags('quanti', 'peptide')

}
\author{
Samuel Wieczorek
}
