% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{separateAdjPval}
\alias{separateAdjPval}
\title{xxx}
\usage{
separateAdjPval(x, pval.threshold = 1.05, method = 1)
}
\arguments{
\item{x}{xxx}

\item{pval.threshold}{xxx}

\item{method}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
NULL

}
\author{
Thomas Burger
}
