% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{setMEC}
\alias{setMEC}
\title{Sets the MEC tag in the metacell}
\usage{
setMEC(qData, conds, df)
}
\arguments{
\item{qData}{xxx}

\item{conds}{xxx}

\item{df}{An object of class \code{MSnSet}}
}
\value{
An instance of class \code{MSnSet}.
}
\description{
Sets the MEC tag in the metacell
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
cols.for.ident <- xxxxx
df <- Biobase::fData(obj)[, cols.for.ident]
setMEC(df, Exp1_R25_pept)

}
\author{
Samuel Wieczorek
}
