% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{setMEC2}
\alias{setMEC2}
\title{Sets the MEC tag in the metacell}
\usage{
setMEC2(conds, df, pattern, level)
}
\arguments{
\item{conds}{xxx}

\item{df}{An object of class \code{MSnSet}}

\item{pattern}{A string to search in metadata that defines if the cell is 
missing or imputed. Available values are `missing` or `imputed`.}

\item{level}{Type of entity/pipeline}

\item{qdata}{xxx}
}
\value{
An instance of class \code{MSnSet}.
}
\description{
This function is based on the metacell dataframe to look for either missing
values (used to update an initial dataset) or imputed values (used when
post processing protein metacell after aggregation)
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept[1:10]
cols.for.ident <- obj@experimentData@other$names_metacell
conds <- Biobase::pData(obj)$Condition
df <- Biobase::fData(obj)[, cols.for.ident]
setMEC2(conds, df, pattern = 'missing', level = 'peptide')

}
\author{
Samuel Wieczorek
}
