% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-pipeline-class.R
\name{show,PipelinePeptide-method}
\alias{show,PipelinePeptide-method}
\title{sdfdsfs}
\usage{
\S4method{show}{PipelinePeptide}(object)
}
\arguments{
\item{object}{xxxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
