% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{show,PipelineTemplate-method}
\alias{show,PipelineTemplate-method}
\title{show function override}
\usage{
\S4method{show}{PipelineTemplate}(object)
}
\arguments{
\item{object}{xxxx}
}
\value{
The value of the slot called designMap
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
