% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{splitAdjacencyMat}
\alias{splitAdjacencyMat}
\title{splits an adjacency matrix into specific and shared}
\usage{
splitAdjacencyMat(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of two adjacency matrices
}
\description{
Method to split an adjacency matrix into specific and shared
}
\examples{
data(Exp1_R25_pept)
obj.pep <- Exp1_R25_pept[seq_len(10)]
protID <- "Protein_group_IDs"
X <- BuildAdjacencyMatrix(obj.pep, protID, FALSE)
ll <- splitAdjacencyMat(X)

}
\author{
Samuel Wieczorek
}
