% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{SumPeptidesSpeedUp}
\alias{SumPeptidesSpeedUp}
\title{Compute the intensity of proteins with the sum of the intensities
of their peptides.}
\usage{
SumPeptidesSpeedUp(matAdj, expr)
}
\arguments{
\item{matAdj}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{expr}{A matrix of intensities of peptides}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins based on the sum of the 
intensities of their peptides.
}
\examples{
data(UPSpep25)
protID <- "Protein.group.IDs"
M <- BuildSparseAdjacencyMatrix(UPSpep25, protID, FALSE)
SumPeptidesSpeedUp(M, Biobase::exprs(UPSpep25))
}
\author{
Alexia Dorffer
}

