% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tabOperator}
\alias{tabOperator}
\title{Selects lines of dataframes in a MSnset file}
\usage{
tabOperator(obj, lineIndices)
}
\arguments{
\item{obj}{A MSnset object}

\item{lineIndices}{The indices of lines to be extracted}
}
\value{
A MSnset object
}
\description{
This function overloads the brackets to select lines of dataframes in a MSnset file. It takes 
into account the slots experimentData / other / OriginOfValues
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
res <- tabOperator(Exp1_R25_pept, c(1:10))
}
\author{
Samuel Wieczorek
}
