% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{test.design.hierarchy}
\alias{test.design.hierarchy}
\title{Check if xxxxxx}
\usage{
test.design.hierarchy(res = NULL, sTab)
}
\arguments{
\item{res}{A list xxx}

\item{sTab}{A data.frame which correspond to xxxxxx}
}
\value{
A list of two items
}
\description{
Check if xxxxxx
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
sTab <- Biobase::pData(Exp1_R25_pept)
test.design.hierarchy(sTab = sTab)

}
\author{
Thomas Burger, Samuel Wieczorek
}
