% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{testAnovaModels}
\alias{testAnovaModels}
\title{xxx}
\usage{
testAnovaModels(aov_fits, test = "Omnibus")
}
\arguments{
\item{aov_fits}{xxx}

\item{test}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
NULL

}
\author{
Thomas Burger
}
