% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{testWithoutNA}
\alias{testWithoutNA}
\title{Test dataset}
\format{An object of class \code{\link{MSnSet}}}
\description{
Partial (small) dataset for unit tests without any
missing values.
}
\keyword{data}

