% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{thresholdpval4fdr}
\alias{thresholdpval4fdr}
\title{xxx}
\usage{
thresholdpval4fdr(x, pval.T, M)
}
\arguments{
\item{x}{xxx}

\item{pval.T}{xxx}

\item{M}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
NULL

}
\author{
Thomas Burger
}
