% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{topMaxUsingPartialSortIndices}
\alias{topMaxUsingPartialSortIndices}
\title{Function to return the indices of the n higher values in the vector}
\usage{
topMaxUsingPartialSortIndices(x, n)
}
\arguments{
\item{x}{A vector of numeric values}

\item{n}{The number of values to be returned}
}
\value{
A vector of the indices of the n highest values
}
\description{
Method to return the indices of the n higher values in the vector
}
\examples{
topMaxUsingPartialSortIndices(c(1:10), 3)
}
\author{
Alexia Dorffer
}

