% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{translatedRandomBeta}
\alias{translatedRandomBeta}
\title{xxxxxxx}
\usage{
translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)
}
\arguments{
\item{n}{xxxxxxxx.}

\item{min}{xxxxxxxxxx}

\item{max}{xxxxxxxxx}

\item{param1}{xxxxxxxxx.}

\item{param2}{xxxxxxxx.}
}
\value{
The object \code{obj} which has been imputed
}
\description{
This method xxxxxxxxxxx
}
\examples{
data(UPSpep25)
translatedRandomBeta(Biobase::exprs(UPSpep25), min=0, max=1)
}
\author{
Thomas Burger
}
