% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioAnalysis.R
\name{univ_AnnotDbPkg}
\alias{univ_AnnotDbPkg}
\title{Returns the totality of ENTREZ ID (gene id) of an OrgDb annotation 
package. Careful : org.Pf.plasmo.db : no ENTREZID but ORF}
\usage{
univ_AnnotDbPkg(orgdb)
}
\arguments{
\item{orgdb}{a Bioconductor OrgDb annotation package}
}
\value{
A vector of ENTREZ ID (totality if the ID for the package)
}
\description{
Returns the universe = totality of ID of an OrgDb annotation package
}
\author{
Florence Combes
}
