% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{updateDataset<-,PipelineTemplate-method}
\alias{updateDataset<-,PipelineTemplate-method}
\title{Set the updateDataset slot value for the object}
\usage{
\S4method{updateDataset}{PipelineTemplate}(obj, name) <- value
}
\arguments{
\item{obj}{xxxx}

\item{name}{xxx}

\item{value}{xxx}
}
\value{
xxxx
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
