% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{UpdateMetacellAfterImputation}
\alias{UpdateMetacellAfterImputation}
\title{Update the cells metadata tags after imputation}
\usage{
UpdateMetacellAfterImputation(obj)
}
\arguments{
\item{obj}{xxx}
}
\value{
xxx
}
\description{
Update the metacell information of missing values that were imputed
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
obj <- Exp1_R25_pept[seq_len(10)]
obj.imp.pov <- wrapper.impute.KNN(obj, K = 3)

}
\author{
Samuel Wieczorek
}
