% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{UPSpepx2}
\alias{UPSpepx2}
\title{UPSpepx2 dataset}
\format{An object of class \code{\link{MSnSet}} related to peptides
quantification. It contains 6 samples divided into two conditions
(25fmol and 50fmol) and 5380 peptides.}
\description{
This dataset contains peptides which result from a controlled relative quantification
proteomics experiment where the commercial Sigma mix UPS1 human proteins
were spiked in a similar yeast lysate in 2 different concentrations
(with a ratio of 2).
}
\keyword{data}

