% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{varianceDistD}
\alias{varianceDistD}
\title{Distribution of variance of proteins}
\usage{
varianceDistD(obj)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}
}
\value{
A density plot
}
\description{
Builds a densityplot of the variance of entities in the exprs() table
of a object. The variance is calculated for each condition (Label) present
in the dataset (see the slot \code{'Label'} in the \code{pData()} table)
}
\examples{
data(UPSprotx2)
varianceDistD(UPSprotx2)
}
\author{
Florence Combes, Samuel Wieczorek
}
\seealso{
\code{\link{densityPlotD}}.
}

