% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-template-class.R
\name{version,PipelineTemplate-method}
\alias{version,PipelineTemplate-method}
\title{Get the version slot value for the object}
\usage{
\S4method{version}{PipelineTemplate}(obj)
}
\arguments{
\item{obj}{xxxx}
}
\value{
The value of the slot version which is a column of the dataset which contains xxxx.
}
\description{
sfklsjhf qsjdhsqd.
}
\examples{
utils::data(Exp1_R25_prot)
}
\author{
Samuel Wieczorek
}
