% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{visualizeClusters}
\alias{visualizeClusters}
\title{Visualize the clusters according to pvalue thresholds}
\usage{
visualizeClusters(
  dat,
  clust_model,
  adjusted_pValues,
  FDR_th = NULL,
  ttl = "",
  subttl = ""
)
}
\arguments{
\item{dat}{the standardize data returned by the function [checkClusterability()]}

\item{clust_model}{the clustering model obtained with dat.}

\item{adjusted_pValues}{vector of the adjusted pvalues obtained for each protein with
a 1-way ANOVA (for example obtained with the function [wrapperClassic1wayAnova()]).}

\item{FDR_th}{the thresholds of FDR pvalues for the coloring of the profiles.
The default (NULL) creates 4 thresholds: 0.001, 0.005, 0.01, 0.05
For the sake of readability, a maximum of 4 values can be specified.}

\item{ttl}{title for the plot.}

\item{subttl}{subtitle for the plot.}
}
\value{
a ggplot object
}
\description{
Visualize the clusters according to pvalue thresholds
}
\examples{
test_anova <- wrapperClassic1wayAnova(fibrose)
vizu <- visualizeClusters(dat = standardized_means,
                          clust_model = km_model,
                          adjusted_pValues = test_anova$P_Value$anova1way,
                          FDR_th = c(0.001,0.005,0.01,0.05),
                          ttl = "Clustering of protein profiles")
                          
}
\author{
Hélène Borges
}
