% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.boxPlotD}
\alias{wrapper.boxPlotD}
\title{Wrapper to the boxplotD function on an object \code{MSnSet}}
\usage{
wrapper.boxPlotD(obj, dataForXAxis = "Condition", ...)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{dataForXAxis}{A vector of strings containing the names of columns 
in \code{pData()} to print conditions on X-axis (Default is "Condition").}

\item{...}{arguments for palette}
}
\value{
A boxplot
}
\description{
This function is a wrapper for using the boxPlotD function with objects of 
class \code{MSnSet}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
types <- c("Condition","Bio.Rep")
wrapper.boxPlotD(Exp1_R25_pept, types)
}
\seealso{
\code{\link{wrapper.densityPlotD}}
}
\author{
Florence Combes, Samuel Wieczorek
}
