% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.compareNormalizationD}
\alias{wrapper.compareNormalizationD}
\title{Builds a plot from a dataframe}
\usage{
wrapper.compareNormalizationD(objBefore, objAfter, labelsForLegend = NULL,
indData2Show = NULL, group2Color = "Condition")
}
\arguments{
\item{objBefore}{A dataframe that contains quantitative data before 
normalization.}

\item{objAfter}{A dataframe that contains quantitative data after 
normalization.}

\item{labelsForLegend}{A vector of the conditions (labels) (one label 
per sample).}

\item{indData2Show}{A vector of the indices of the columns to show in the 
plot. The indices are those of indices of 
the columns int the data.frame qDataBefore.}

\item{group2Color}{A string that indicates how to color the replicates: 
one color per condition (value "Condition") or one 
color per replicate (value "Replicate"). Default value is by Condition.}
}
\value{
A plot
}
\description{
Wrapper to the function that plot to compare the quantitative proteomics 
data before and after normalization
}
\examples{
data(UPSpep25)
labels <- pData(UPSpep25)[,"Label"]
objAfter <- wrapper.normalizeD(UPSpep25, "Median Centering", 
"within conditions")
wrapper.compareNormalizationD(UPSpep25, objAfter, labels)
}
\author{
Samuel Wieczorek
}

