% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.CVDistD}
\alias{wrapper.CVDistD}
\title{Distribution of CV of entities}
\usage{
wrapper.CVDistD(obj)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}
}
\value{
A density plot
}
\description{
Builds a densityplot of the CV of entities in the exprs() table
of an object \code{\link{MSnSet}}. The variance is calculated for each 
condition (Label) present
in the dataset (see the slot \code{'Label'} in the \code{pData()} table).
}
\examples{
data(UPSpep25)
wrapper.CVDistD(UPSpep25)
}
\seealso{
\code{\link{wrapper.densityPlotD}}
}
\author{
Alexia Dorffer
}
