% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_density_CV.R
\name{wrapper.CVDistD_HC}
\alias{wrapper.CVDistD_HC}
\title{Distribution of CV of entities}
\usage{
wrapper.CVDistD_HC(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}}

\item{...}{arguments for palette.}
}
\value{
A density plot
}
\description{
Builds a densityplot of the CV of entities in the 
Biobase::exprs() table. of an object \code{MSnSet}. The variance is 
calculated for each condition present in the dataset (see the slot 
\code{'Condition'} in the \code{Biobase::pData()} table).
}
\examples{
data(Exp1_R25_pept)
wrapper.CVDistD_HC(Exp1_R25_pept)

}
\author{
Samuel Wieczorek
}
