% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{wrapper.dapar.impute.mi}
\alias{wrapper.dapar.impute.mi}
\title{Missing values imputation using the LSimpute algorithm.}
\usage{
wrapper.dapar.impute.mi(obj, nb.iter = 3, nknn = 15, selec = 600,
  siz = 500, weight = 1, ind.comp = 1, progress.bar = TRUE,
  x.min = 10, x.max = 35, x.step.mod = 300, x.step.pi = 300,
  nb.rei = 100, method = 4, gridsize = 300, q = 0.95, q.min = 0,
  q.norm = 3, eps = 0, methodi = "slsa", lapala = TRUE)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{selec}{An integer which corresponds to the \code{selec} parameter of the
function \code{impute.slsa}.}
}
\value{
The \code{exprs(obj)} matrix with imputed values instead of missing values.
}
\description{
This method is a wrapper to the function \code{impute.mi} of the package \code{imp4p} adapted to
an object of class \code{\link{MSnSet}}.
}
\examples{
data(UPSpep25)
dat <- mvFilter(UPSpep25, type="allCond", th = 1)
dat <- wrapper.dapar.impute.mi(dat)
}
\author{
Samuel Wieczorek
}
