% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.densityPlotD}
\alias{wrapper.densityPlotD}
\title{Builds a densityplot from an object of class \code{\link{MSnSet}}}
\usage{
wrapper.densityPlotD(obj, lab2Show = NULL, highLightLabel = NULL)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{lab2Show}{A vector of labels to show in densityplot.}

\item{highLightLabel}{The name of the Label to highlight
in the density plot.}
}
\value{
A density plot
}
\description{
This function is a wrapper for using the densityPlotD function with objects of class \code{\link{MSnSet}}
}
\examples{
data(UPSprotx2)
wrapper.densityPlotD(UPSprotx2)
}
\author{
Alexia Dorffer
}
\seealso{
\code{\link{wrapper.boxPlotD}}, \code{\link{wrapper.varianceDistD}}
}

