% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.densityPlotD_HC}
\alias{wrapper.densityPlotD_HC}
\title{Builds a densityplot from an object of class \code{MSnSet}}
\usage{
wrapper.densityPlotD_HC(obj, condsForLegend = NULL, indData2Show = NULL,
  group2Color = "Condition")
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{condsForLegend}{A vector of conds to show in densityplot.}

\item{indData2Show}{A vector of the indices of the columns to show in 
the plot. The indices are those of indices of the columns int the data
frame qDataBefore in the density plot.}

\item{group2Color}{A string that indicates how to color the replicates: 
one color per condition (value "Condition") or one color per replicate
(value "Replicate"). Default value is by Condition.}
}
\value{
A density plot
}
\description{
This function is a wrapper for using the densityPlotD function with 
objects of class \code{MSnSet}. Same as the function \code{\link{wrapper.densityPlotD}}
but uses the package \code{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
wrapper.densityPlotD_HC(Exp1_R25_pept, conds)
}
\seealso{
\code{\link{wrapper.boxPlotD}}, 
\code{\link{wrapper.CVDistD}}
}
\author{
Samuel Wieczorek
}
