% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{wrapper.diffAnaLimma}
\alias{wrapper.diffAnaLimma}
\title{Performs differential analysis on
an MSnSet object, calling the \code{limma} package functions}
\usage{
wrapper.diffAnaLimma(obj, condition1, condition2)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{condition1}{A vector that contains the names of the conditions 
considered as condition 1.}

\item{condition2}{A vector that contains the names of the conditions 
considered as condition 2.}
}
\value{
A dataframe as returned by the \code{limma} package
}
\description{
Method to perform differential analysis on
a \code{MSnSet} object (calls the \code{limma} package function).
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
condition1 <- '25fmol'
condition2 <- '10fmol'
wrapper.diffAnaLimma(Exp1_R25_pept[1:1000], condition1, condition2)
}
\author{
Alexia Dorffer
}
