% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{wrapper.diffAnaWelch}
\alias{wrapper.diffAnaWelch}
\title{Performs a differential analysis on a \code{\link{MSnSet}} object
using the Welch t-test}
\usage{
wrapper.diffAnaWelch(obj, condition1, condition2)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{condition1}{A vector containing the names of the conditions 
considered as condition 1.}

\item{condition2}{A vector containing the names of the conditions 
considered as condition 2.}
}
\value{
A dataframe with two slots : P.Value (for the p-value) and logFC
(the log of the Fold Change).
}
\description{
Computes differential analysis on
a \code{\link{MSnSet}} object, using the Welch t-test
(\code{\link{t.test}{stats}}).
}
\examples{
library(DAPARdata)
data(UPSprotx2)
condition1 <- '10fmol'
condition2 <- '5fmol'
wrapper.diffAnaWelch(UPSprotx2, condition1, condition2)
}
\author{
Alexia Dorffer
}

