% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{wrapper.identifyMCAR_MNAR}
\alias{wrapper.identifyMCAR_MNAR}
\title{Estimation of a mixture model of MCAR and MNAR values.}
\usage{
wrapper.identifyMCAR_MNAR(obj, dat.slsa)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{dat.slsa}{This parameter corresponds to the parameter \code{tab.imp} in the function \code{estim.mix}.}
}
\value{
A list. For details, please refer to the value section of the help for the function \code{estim.mix}.
}
\description{
This method is a wrapper to the function \code{estim.mix} of the package \code{imp4p} adapted to 
an object of class \code{\link{MSnSet}}.
}
\examples{
data(UPSpep25)
dat <- mvFilter(UPSpep25, type="allCond", th = 1)
dat.slsa <- wrapper.impute.slsa(dat)
proba <- wrapper.identifyMCAR_MNAR(dat,dat.slsa)
}
\author{
Samuel Wieczorek
}

