% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{wrapper.impute.detQuant}
\alias{wrapper.impute.detQuant}
\title{xxxxxxxxx}
\usage{
wrapper.impute.detQuant(obj, qval = 0.025, factor = 1)
}
\arguments{
\item{obj}{xxxxxxxxx}

\item{qval}{xxxxxxxxx}

\item{factor}{xxxxxxxxx.}
}
\value{
xxxxxxxxx
}
\description{
This method is xxxxxxxx
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.impute.detQuant(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
