% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_PeptideLevel.R
\name{wrapper.impute.mle}
\alias{wrapper.impute.mle}
\title{Imputation of peptides having no values in a biological condition.}
\usage{
wrapper.impute.mle(obj = NULL, na.type = NULL)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{na.type}{A string which indicates the type of missing values to impute. 
Available values are: `NA` (for both POV and MEC).}
}
\value{
The \code{exprs(obj)} matrix with imputed values instead of missing values.
}
\description{
This method is a wrapper to the function \code{impute.mle} of the package
\code{imp4p} adapted to an object of class \code{MSnSet}.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
dat <- mvFilter(Exp1_R25_pept[1:1000,], type="AllCond", th = 1)
dat <- wrapper.impute.mle(dat)

}
\author{
Samuel Wieczorek
}
