% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{wrapper.imputeImp4p}
\alias{wrapper.imputeImp4p}
\title{Multiple imputation form a matrix of probabilities of being MCAR for each missing value..}
\usage{
wrapper.imputeImp4p(obj, dat.slsa, proba, nb.iter = 1, selec = 100,
  ind.comp = 0)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{dat.slsa}{This parameter corresponds to the parameter \code{tab.imp} in the function \code{mi.mix}.}

\item{proba}{This parameter corresponds 
to the parameter \code{proba.MCAR} in the function \code{mi.mix}.}

\item{nb.iter}{The number of iterations used for the multiple imputation method. 
This parameter corresponds to the parameter \code{nb.iter} in the function \code{mi.mix}.}

\item{selec}{This parameter corresponds to the parameter \code{selec} in the function \code{mi.mix}.}

\item{ind.comp}{This parameter corresponds to the parameter \code{ind.comp} in the function \code{mi.mix}.}
}
\value{
The \code{obj}  with imputed values instead of missing values.
}
\description{
This method is a wrapper to the function \code{mi.mix} of the package \code{imp4p} adapted to 
an object of class \code{\link{MSnSet}}.
}
\examples{
data(UPSpep25)
dat <- mvFilter(UPSpep25, type="allCond", th = 1)
dat.slsa <- wrapper.impute.slsa(dat)
proba <- wrapper.identifyMCAR_MNAR(dat,dat.slsa)
wrapper.imputeImp4p(dat, dat.slsa, proba)
}
\author{
Samuel Wieczorek
}

