% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{wrapper.limmaCompleteTest}
\alias{wrapper.limmaCompleteTest}
\title{Computation of the hierarchical design matrix : 2-level case : Bio-Tech or 
Bio-Analytical or Tech-Analytical.}
\usage{
wrapper.limmaCompleteTest(obj, design)
}
\arguments{
\item{obj}{An object of class \code{MSnSet} with no missing values}

\item{design}{An integer that reflects the type of comparisons. Available values are 1 (One vs One) or (One vs All)}
}
\value{
xxxxxxxxxx
}
\description{
This function is a wrapper of \code{limmaCompleteTest} for objects of class \code{MSnSet}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[1:1000]
lapala <- findMECBlock(obj)
obj <- wrapper.impute.detQuant(obj)
obj <- reIntroduceMEC(obj, lapala)
obj <- wrapper.impute.detQuant(obj)
limma <- wrapper.limmaCompleteTest(obj, 1)
}
\author{
Samuel Wieczorek
}
