% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{wrapper.mvHisto}
\alias{wrapper.mvHisto}
\title{Histogram of missing values from a \code{\link{MSnSet}} object}
\usage{
wrapper.mvHisto(obj, indLegend = "auto", showValues = FALSE)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{indLegend}{The indices of the column name's in \code{pData()} tab.}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A histogram
}
\description{
This method plots from a \code{\link{MSnSet}} object a histogram of missing values.
}
\examples{
library(DAPARdata)
data(UPSpep25)
wrapper.mvHisto(UPSpep25, showValues=TRUE)
}
\author{
Alexia Dorffer
}

