% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{wrapper.mvImputation}
\alias{wrapper.mvImputation}
\title{Missing values imputation from a \code{MSnSet} object}
\usage{
wrapper.mvImputation(obj, method)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{method}{The imputation method to be used.
Choices are QRILC, KNN, BPCA and MLE.}
}
\value{
The object \code{obj} which has been imputed
}
\description{
This method is a wrapper to the \code{imputeLCMD} package adapted to 
objects of class \code{MSnSet}.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.mvImputation(Exp1_R25_pept[1:1000], "QRILC")
}
\author{
Alexia Dorffer
}
