% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{wrapper.mvPerLinesHisto}
\alias{wrapper.mvPerLinesHisto}
\title{Histogram of missing values per lines from an object \code{\link{MSnSet}}}
\usage{
wrapper.mvPerLinesHisto(obj, indLegend = "auto", showValues = FALSE)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{indLegend}{The indice of the column name's in \code{pData()} tab .}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A histogram
}
\description{
This method is a wrapper to plots from a \code{\link{MSnSet}} object a histogram which represents the distribution of the 
number of missing values (NA) per lines (ie proteins).
}
\examples{
library(DAPARdata)
data(UPSpep25)
wrapper.mvPerLinesHisto(UPSpep25)
}
\author{
Alexia Dorffer
}

