% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{wrapper.mvTypePlot}
\alias{wrapper.mvTypePlot}
\title{Distribution of missing values with respect to intensity values 
from a \code{MSnSet} object}
\usage{
wrapper.mvTypePlot(obj, threshold = 0)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{threshold}{An integer for the intensity that delimits MNAR and 
MCAR missing values.}
}
\value{
A scatter plot
}
\description{
This method plots a scatter plot which represents the distribution of
missing values.
The colors correspond to the different conditions (slot Label in in the
dataset of class \code{MSnSet}).
The x-axis represent the mean of intensity for one condition and one
entity in the dataset (i. e. a protein) 
whereas the y-axis count the number of missing values for this entity
and the considered condition.
The data have been jittered for an easier vizualisation.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.mvTypePlot(Exp1_R25_pept)
}
\author{
Florence Combes, Samuel Wieczorek
}
