% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{wrapper.normalizeD}
\alias{wrapper.normalizeD}
\title{Normalisation}
\usage{
wrapper.normalizeD(obj, family, method)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{family}{One of the following : Global Rescaling, 
Median Centering, Mean Centering, Mean Centering Scaling.}

\item{method}{"Overall" or "within conditions".}
}
\value{
An instance of class \code{\link{MSnSet}} 
where the quantitative data in the \code{exprs()} tab
has been normalized.
}
\description{
Provides several methods to normalize quantitative data from
a \code{\link{MSnSet}} object.
They are organized in four main families : Strong Rescaling, 
Median Centering, Mean Centering, Mean CenteringScaling.
For the first family, two sub-categories are available : the sum by columns
and the quantiles method.
For the three other families, two categories are available : 
"Overall" which means that the value for each protein 
(ie line in the expression data tab) is computed over all the samples ;
"within conditions" which means that the value for each protein 
(ie line in the \code{exprs()} data tab) is computed condition
by condition.
}
\examples{
library(DAPARdata)
data(UPSpep25)
wrapper.normalizeD(UPSpep25, "Median Centering", "within conditions")
}
\author{
Alexia Dorffer
}

