% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_pca.R
\name{wrapper.pca}
\alias{wrapper.pca}
\title{Compute the PCA}
\usage{
wrapper.pca(obj, var.scaling = TRUE, ncp = NULL)
}
\arguments{
\item{obj}{xxx}

\item{var.scaling}{The dimensions to plot}

\item{ncp}{xxxx}
}
\value{
A xxxxxx
}
\description{
Compute the PCA
}
\examples{
utils::data(Exp1_R25_prot, package='DAPARdata')
obj <- Exp1_R25_prot[1:1000]
level <- obj@experimentData@other$typeOfData
metacell.mask <- match.metacell(GetMetacell(obj), 'missing', level)
indices <- GetIndices_WholeMatrix(metacell.mask, op='>=', th=1)
obj <- MetaCellFiltering(obj, indices, cmd='delete')
res.pca <- wrapper.pca(obj$new)


}
\author{
Samuel Wieczorek
}
