% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.varianceDistD}
\alias{wrapper.varianceDistD}
\title{Distribution of variance of proteins}
\usage{
wrapper.varianceDistD(obj)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}
}
\value{
A density plot
}
\description{
Builds a densityplot of the variance of entities in the exprs() table
of an object \code{\link{MSnSet}}. The variance is calculated for each condition (Label) present
in the dataset (see the slot \code{'Label'} in the \code{pData()} table).
}
\examples{
library(DAPARdata)
data(UPSprot25)
wrapper.varianceDistD(UPSprot25)
}
\author{
Alexia Dorffer
}
\seealso{
\code{\link{wrapper.densityPlotD}}
}

