% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{writeMSnsetToCSV}
\alias{writeMSnsetToCSV}
\title{Exports a MSnset dataset into a zip archive containing three
zipped CSV files.}
\usage{
writeMSnsetToCSV(obj, fname)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{fname}{The name of the archive file.}
}
\value{
A compressed file
}
\description{
Exports a MSnset dataset into a zip archive containing three
zipped CSV files.
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
obj <- Exp1_R25_pept[seq_len(10)]
writeMSnsetToCSV(obj, "foo")

}
\author{
Samuel Wieczorek
}
