% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{AggregateMetacell}
\alias{AggregateMetacell}
\title{Symbolic product of matrices}
\usage{
AggregateMetacell(X, obj.pep)
}
\arguments{
\item{X}{An adjacency matrix between peptides and proteins}

\item{obj.pep}{A dataframe of the cell metadata for peptides}
}
\value{
xxxx
}
\description{
Execute a product two matrices: the first is an adjacency one  while the
second if a simple dataframe
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj.pep <- Exp1_R25_pept[1:100]
protID <- "Protein_group_IDs"
X <- BuildAdjacencyMatrix(obj.pep, protID, FALSE)
agg.meta <- AggregateMetacell(X, obj.pep)

}
\author{
Samuel Wieczorek
}
