% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{applyAnovasOnProteins}
\alias{applyAnovasOnProteins}
\title{xxx}
\usage{
applyAnovasOnProteins(obj)
}
\arguments{
\item{obj}{xxx
'}
}
\value{
xxx
}
\description{
xxx
}
\examples{
NULL

}
\author{
Thomas Burger
}
