% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{boxPlotD}
\alias{boxPlotD}
\title{Builds a boxplot from an object of class \code{\link{MSnSet}}}
\usage{
boxPlotD(obj, Xaxis = "Label")
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{Xaxis}{A vector containing the indices of columns
in \code{pData()} to use as X-axis (Default is "Label").}
}
\value{
A boxplot
}
\description{
Boxplot for quantitative proteomics data
}
\examples{
data(UPSprotx2)
boxPlotD(UPSprotx2)
}
\author{
Florence Combes, Samuel Wieczorek
}
\seealso{
\code{\link{densityPlotD}}
}

