% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{boxPlotD}
\alias{boxPlotD}
\title{Builds a boxplot from a dataframe}
\usage{
boxPlotD(qData, dataForXAxis = NULL, labels = NULL,
  group2Color = "Condition")
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{dataForXAxis}{A vector containing the types of replicates 
to use as X-axis. Available values are: Label, Analyt.Rep,
Bio.Rep and Tech.Rep. Default is "Label".}

\item{labels}{A vector of the conditions (labels) (one label per sample).}

\item{group2Color}{A string that indicates how to color the replicates: 
one color per condition (value "Condition") or one 
color per replicate (value "Replicate"). Default value is by Condition.}
}
\value{
A boxplot
}
\description{
Boxplot for quantitative proteomics data
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
types <- c("Label","Analyt.Rep")
dataForXAxis <- Biobase::pData(Exp1_R25_pept)[,types]
labels <- Biobase::pData(Exp1_R25_pept)[,"Label"]
boxPlotD(qData, dataForXAxis, labels)
}
\seealso{
\code{\link{densityPlotD}}
}
\author{
Florence Combes, Samuel Wieczorek
}
