% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{boxPlotD_HC}
\alias{boxPlotD_HC}
\title{Builds a boxplot from a dataframe using the library \code{highcharter}}
\usage{
boxPlotD_HC(qData, dataForXAxis = "Condition", conds = NULL,
  palette = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{dataForXAxis}{A vector containing the types of replicates 
to use as X-axis. Available values are: Condition, Bio.Rep,
Bio.Rep and Tech.Rep. Default is "Condition".}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}
}
\value{
A boxplot
}
\description{
Boxplot for quantitative proteomics data using the library \code{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
types <- c("Condition","Bio.Rep")
dataForXAxis <- Biobase::pData(Exp1_R25_pept)[,types]
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
boxPlotD_HC(qData, dataForXAxis, conds)
}
\seealso{
\code{\link{densityPlotD_HC}}
}
\author{
Samuel Wieczorek
}
