% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{boxPlotD_HC}
\alias{boxPlotD_HC}
\title{Builds a boxplot from a dataframe using the library \code{\link{highcharter}}}
\usage{
boxPlotD_HC(qData, dataForXAxis = "Label", labels = NULL,
  group2Color = "Condition")
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{dataForXAxis}{A vector containing the types of replicates 
to use as X-axis. Available values are: Label, Analyt.Rep,
Bio.Rep and Tech.Rep. Default is "Label".}

\item{labels}{A vector of the conditions (labels) (one label per sample).}

\item{group2Color}{A string that indicates how to color the replicates: 
one color per condition (value "Condition") or one 
color per replicate (value "Replicate"). Default value is by Condition.}
}
\value{
A boxplot
}
\description{
Boxplot for quantitative proteomics data using the library \code{\link{highcharter}}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
types <- c("Label","Analyt.Rep")
dataForXAxis <- Biobase::pData(Exp1_R25_pept)[,types]
labels <- Biobase::pData(Exp1_R25_pept)[,"Label"]
boxPlotD_HC(qData, dataForXAxis, labels)
}
\seealso{
\code{\link{densityPlotD_HC}}
}
\author{
Samuel Wieczorek
}
