% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{BuildColumnToProteinDataset}
\alias{BuildColumnToProteinDataset}
\title{creates a column for the protein dataset after agregation by
 using the previous peptide dataset.}
\usage{
BuildColumnToProteinDataset(peptideData, matAdj, columnName)
}
\arguments{
\item{peptideData}{A data.frame of meta data of peptides. It is the fData of the 
MSnset object.}

\item{matAdj}{The adjacency matrix used to agregate the peptides data.}

\item{columnName}{The name of the column in fData(peptides_MSnset) that the user
wants to keep in the new protein data.frame.}
}
\value{
A vector
}
\description{
This function creates a column for the protein dataset after agregation by using the previous peptide dataset.
}
\examples{
library(DAPARdata)
data(UPSpepx2)
protID <- "Protein.group.IDs"
M <- BuildAdjacencyMatrix(UPSpepx2, protID, FALSE)
data <- fData(UPSpepx2)
name <- "organism"
BuildColumnToProteinDataset(data, M, name )
}
\author{
Samuel Wieczorek
}

