% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BuildPalette}
\alias{BuildPalette}
\title{Builds a complete color palette for the conditions given in argument}
\usage{
BuildPalette(conds, base_palette = NULL)
}
\arguments{
\item{conds}{The extended vector of samples conditions}

\item{base_palette}{The basic color (HEX code) used to build the complete palette. This vector have the same length as unique(conds).
Default base palette is 'Dark2' from the package RColorBrewer.}
}
\value{
A vector composed of HEX color code for the conditions
}
\description{
xxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
conditions <- Biobase::pData(Exp1_R25_pept)$Condition
BuildPalette(conditions, c('AAAAAA', 'BBBBBB'))

}
\author{
Samuel Wieczorek
}
