% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{classic1wayAnova}
\alias{classic1wayAnova}
\title{Function to perform a One-way Anova statistical test on a MsnBase 
dataset}
\usage{
classic1wayAnova(current_line, conditions)
}
\arguments{
\item{current_line}{The line currently treated from the quantitative data to
perform the ANOVA}

\item{conditions}{The conditions represent the different classes of the
studied factor}
}
\value{
A named vector containing all the different values of the aov model
}
\description{
Function to perform a One-way Anova statistical test on a MsnBase 
dataset
}
\examples{
data(Exp1_R25_prot)
obj <- Exp1_R25_prot[seq_len(1000)]
level <- 'protein'
metacell.mask <- match.metacell(GetMetacell(obj), "missing", level)
indices <- GetIndices_WholeMatrix(metacell.mask, op = ">=", th = 1)
obj <- MetaCellFiltering(obj, indices, cmd = "delete")
anova_tests <- t(apply(Biobase::exprs(obj$new), 1, classic1wayAnova,
    conditions = as.factor(Biobase::pData(obj$new)$Condition)
))

}
\author{
Hélène Borges
}
