% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{compareNormalizationD}
\alias{compareNormalizationD}
\title{Builds a plot from a dataframe}
\usage{
compareNormalizationD(qDataBefore, qDataAfter, dataForXAxis = NULL,
  labels = NULL)
}
\arguments{
\item{qDataBefore}{A dataframe that contains quantitative data before normalization.}

\item{qDataAfter}{A dataframe that contains quantitative data after normalization.}

\item{dataForXAxis}{A vector containing the types of replicates 
to use as X-axis. Available values are: Label, Analyt.Rep,
Bio.Rep and Tech.Rep. Default is "Label".}

\item{labels}{A vector of the conditions (labels) (one label per sample).}
}
\value{
A plot
}
\description{
Plot to compare the quantitative proteomics data before and after normalization
}
\examples{
data(UPSprotx2)
qDataBefore <- exprs(UPSprotx2)
labels <- pData(UPSprotx2)[,"Label"]
qDataAfter <- normalizeD(qDataBefore, labels, "Median Centering", "within conditions")
types <- c("Label","Analyt.Rep")
dataForXAxis <- pData(UPSprotx2)[,types]
compareNormalizationD(qDataBefore, qDataAfter, dataForXAxis, labels)
}
\author{
Florence Combes, Samuel Wieczorek
}

