% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{compareNormalizationD_HC}
\alias{compareNormalizationD_HC}
\title{Builds a plot from a dataframe. Same as compareNormalizationD but 
uses the Highcharter library}
\usage{
compareNormalizationD_HC(qDataBefore, qDataAfter, labelsForLegend = NULL,
  indData2Show = NULL, group2Color = "Condition")
}
\arguments{
\item{qDataBefore}{A dataframe that contains quantitative data before 
normalization.}

\item{qDataAfter}{A dataframe that contains quantitative data after 
normalization.}

\item{labelsForLegend}{A vector of the conditions (labels) (one label 
per sample).}

\item{indData2Show}{A vector of the indices of the columns to show in 
the plot. The indices are those of indices of 
the columns int the data.frame qDataBefore.}

\item{group2Color}{A string that indicates how to color the replicates:
one color per condition (value "Condition") or one 
color per replicate (value "Replicate"). Default value is by Condition.}
}
\value{
A plot
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qDataBefore <- Biobase::exprs(Exp1_R25_pept)
labels <- Biobase::pData(Exp1_R25_pept)[,"Label"]
qDataAfter <- normalizeD(qDataBefore,labels,"Median Centering",
"within conditions")
compareNormalizationD_HC(qDataBefore, qDataAfter, labels)
}
\author{
Samuel Wieczorek
}
