% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{corrMatrixD}
\alias{corrMatrixD}
\title{Displays a correlation matrix of the quantitative data of the
\code{exprs()} table.}
\usage{
corrMatrixD(qData, samplesData, indLegend = NULL, gradientRate = 5)
}
\arguments{
\item{qData}{A dataframe of quantitative data.}

\item{samplesData}{A dataframe where lines correspond to samples and columns to the meta-data for those samples.}

\item{indLegend}{A vector of indices in the columns of
the samplesData table choosen for the labels in the axes.}

\item{rate}{The rate parameter to control de exponential law for the gradient of colors}
}
\value{
A colored correlation matrix
}
\description{
Correlation matrix based on a \code{\link{MSnSet}} object
}
\examples{
data(UPSprotx2)
qData <- exprs(UPSprotx2)
samplesData <- pData(UPSprotx2)
corrMatrixD(qData, samplesData)
}
\author{
Florence Combes, Samuel Wieczorek
}

