% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{corrMatrixD}
\alias{corrMatrixD}
\title{Displays a correlation matrix of the quantitative data of the
\code{exprs()} table.}
\usage{
corrMatrixD(qData, samplesData, gradientRate = 5)
}
\arguments{
\item{qData}{A dataframe of quantitative data.}

\item{samplesData}{A dataframe where lines correspond to samples and 
columns to the meta-data for those samples.}

\item{gradientRate}{The rate parameter to control the exponential law for 
the gradient of colors}
}
\value{
A colored correlation matrix
}
\description{
Correlation matrix based on a \code{\link{MSnSet}} object
}
\examples{
data(UPSpep25)
qData <- Biobase::exprs(UPSpep25)
samplesData <- Biobase::pData(UPSpep25)
corrMatrixD(qData, samplesData)
}
\author{
Florence Combes, Samuel Wieczorek
}
