% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{corrMatrixD_HC}
\alias{corrMatrixD_HC}
\title{Displays a correlation matrix of the quantitative data of the
\code{exprs()} table.}
\usage{
corrMatrixD_HC(object, samplesData = NULL, minColor = 0.5)
}
\arguments{
\item{samplesData}{A dataframe where lines correspond to samples and 
columns to the meta-data for those samples.}

\item{qData}{A dataframe of quantitative data.}

\item{gradientRate}{The rate parameter to control the exponential law for 
the gradient of colors}
}
\value{
A colored correlation matrix
}
\description{
Correlation matrix based on a \code{\link{MSnSet}} object. Same as the 
function \link{corrMatrixD} but uses the package \CRANpkg{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
samplesData <- Biobase::pData(Exp1_R25_pept)
corrMatrixD_HC(cor(qData,use = 'pairwise.complete.obs'),samplesData)
}
\author{
Samuel Wieczorek
}
