% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{CVDistD_HC}
\alias{CVDistD_HC}
\title{Distribution of CV of entities}
\usage{
CVDistD_HC(qData, labels = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{labels}{A vector of the conditions (labels) (one label per sample).}
}
\value{
A density plot
}
\description{
Builds a densityplot of the CV of entities in the exprs() table
of a object. The CV is calculated for each condition (Label) present
in the dataset (see the slot \code{'Label'} in the \code{pData()} table)
Same as the function \code{CVDistD} but uses the package \CRANpkg{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
labels <- Biobase::pData(Exp1_R25_pept)[,"Label"]
CVDistD_HC(Biobase::exprs(Exp1_R25_pept), labels)
}
\seealso{
\code{\link{densityPlotD}}.
}
\author{
Samuel Wieczorek
}
