% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{densityPlotD}
\alias{densityPlotD}
\title{Builds a densityplot from a dataframe}
\usage{
densityPlotD(qData, labels = NULL, lab2Show = NULL, highLightLabel = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{labels}{A vector of the conditions (labels) (one label per sample).}

\item{lab2Show}{A vector of labels to show in densityplot. If NULL, then all labels are displayed.}

\item{highLightLabel}{The name of the Label to highlight
in the density plot.}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples.
}
\examples{
data(UPSprotx2)
qData <- exprs(UPSprotx2)
labels <- lab2Show <- pData(UPSprotx2)[,"Label"]
densityPlotD(qData, labels, lab2Show)
}
\author{
Florence Combes, Samuel Wieczorek
}
\seealso{
\code{\link{boxPlotD}}, \code{\link{varianceDistD}}
}

