% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{densityPlotD}
\alias{densityPlotD}
\title{Builds a densityplot from a dataframe}
\usage{
densityPlotD(qData, condsForLegend = NULL, indData2Show = NULL,
  group2Color = "Condition")
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{condsForLegend}{A vector of the conditions (one condition per sample).}

\item{indData2Show}{A vector of indices to show in densityplot. If NULL, 
then all conds are displayed.}

\item{group2Color}{A string that indicates how to color the replicates: 
one color per condition (value "Condition") or one 
color per replicate (value "Replicate"). Default value is by Condition.}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
conds <- lab2Show <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
densityPlotD(qData, conds)
}
\seealso{
\code{\link{boxPlotD}}, \code{\link{CVDistD}}
}
\author{
Florence Combes, Samuel Wieczorek
}
