% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_density.R
\name{densityPlotD_HC}
\alias{densityPlotD_HC}
\title{Builds a densityplot from a dataframe}
\usage{
densityPlotD_HC(obj, legend = NULL, pal = NULL)
}
\arguments{
\item{obj}{xxx}

\item{legend}{A vector of the conditions (one condition
per sample).}

\item{pal}{xxx}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples.
}
\examples{
data(Exp1_R25_pept)
densityPlotD_HC(Exp1_R25_pept)
conds <- Biobase::pData(Exp1_R25_pept)$Condition
pal <- ExtendPalette(2, "Dark2")
densityPlotD_HC(Exp1_R25_pept, pal = pal)

}
\author{
Samuel Wieczorek
}
