% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{densityPlotD_HC}
\alias{densityPlotD_HC}
\title{Builds a densityplot from a dataframe}
\usage{
densityPlotD_HC(qData, condsForLegend = NULL, indData2Show = NULL,
  palette = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{condsForLegend}{A vector of the conditions (one condition 
per sample).}

\item{indData2Show}{A vector of indices to show in densityplot. If NULL, 
then all conditions are displayed.}

\item{palette}{xxx}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples. Same as the function \code{\link{densityPlotD}}
but uses the package \code{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
conds <- lab2Show <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
densityPlotD_HC(qData, conds)
}
\seealso{
\code{\link{boxPlotD}}, \code{\link{CVDistD}}
}
\author{
Samuel Wieczorek
}
