% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{densityPlotD_HC}
\alias{densityPlotD_HC}
\title{Builds a densityplot from a dataframe}
\usage{
densityPlotD_HC(qData, labelsForLegend = NULL, indData2Show = NULL,
  group2Color = "Condition")
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{labelsForLegend}{A vector of the conditions (labels) (one label 
per sample).}

\item{indData2Show}{A vector of indices to show in densityplot. If NULL, 
then all labels are displayed.}

\item{group2Color}{A string that indicates how to color the replicates: 
one color per condition (value "Condition") or one 
color per replicate (value "Replicate"). Default value is by Condition.}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples. Same as the function \code{\link{densityPlotD}}
but uses the package \code{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
labels <- lab2Show <- Biobase::pData(Exp1_R25_pept)[,"Label"]
densityPlotD_HC(qData, labels)
}
\seealso{
\code{\link{boxPlotD}}, \code{\link{CVDistD}}
}
\author{
Samuel Wieczorek
}
