% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAna}
\alias{diffAna}
\title{This function performs a differential analysis on an MSnSet
object (adapted from \code{\link{limma}})}
\usage{
diffAna(qData, design)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{design}{The design matrix as described in the limma package 
documentation}
}
\value{
A dataframe with the p-value and log(Fold Change) associated 
to each element (peptide/protein)
}
\description{
Performs a differential analysis on an \code{\link{MSnSet}} object,
based on \code{\link{limma}} functions.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept[1:1000])
design <- cbind(cond1=1, 
cond2 = rep(0,nrow(Biobase::pData(Exp1_R25_pept[1:1000]))))
rownames(design) <- rownames(Biobase::pData(Exp1_R25_pept[1:1000]))
labels <- Biobase::pData(Exp1_R25_pept[1:1000])[,"Label"]
indices <- getIndicesConditions(labels, "25fmol", "10fmol")
design[indices$iCond2,2] <- 1
diffAna(qData, design)
}
\author{
Florence Combes, Samuel Wieczorek
}
