% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaComputeFDR}
\alias{diffAnaComputeFDR}
\title{Computes the FDR corresponding to the p-values of the
differential analysis.}
\usage{
diffAnaComputeFDR(data, threshold_PVal = 0, threshold_LogFC = 0)
}
\arguments{
\item{data}{The result of the differential analysis processed
by \code{\link{diffAna}}}

\item{threshold_PVal}{The threshold on p-pvalue to
distinguish between differential and non-differential data}

\item{threshold_LogFC}{The threshold on log(Fold Change) to
distinguish between differential and non-differential data}
}
\value{
The computed FDR value (floating number)
}
\description{
This function returns the FDR corresponding to the p-values of the
differential analysis.
The FDR is computed with the function \code{p.adjust}\{stats\}, with the
\code{BH} correction method (Benjamini & Hochberg (1995)).
}
\examples{
data(UPSprotx2)
obj <- mvImputation(UPSprotx2, "QRILC")
condition1 <- '10fmol'
condition2 <- '5fmol'
limma <- diffAnaLimma(obj, condition1, condition2)
diffAnaComputeFDR(limma)
}
\author{
Alexia Dorffer
}

