% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaComputeFDR}
\alias{diffAnaComputeFDR}
\title{Computes the FDR corresponding to the p-values of the 
differential analysis using}
\usage{
diffAnaComputeFDR(data, threshold_PVal = 0, threshold_LogFC = 0,
  pi0Method = 1)
}
\arguments{
\item{data}{The result of the differential analysis processed 
by \code{\link{diffAna}}}

\item{threshold_PVal}{The threshold on p-pvalue to
distinguish between differential and non-differential data}

\item{threshold_LogFC}{The threshold on log(Fold Change) to
distinguish between differential and non-differential data}

\item{pi0Method}{The parameter pi0.method of the method adjust.p 
in the package \code{cp4p}}
}
\value{
The computed FDR value (floating number)
}
\description{
This function is a wrappper to the function adjust.p from the
cp4p package. It returns the FDR corresponding to the p-values of the 
differential analysis.
The FDR is computed with the function \code{p.adjust}\{stats\}..
}
\examples{
data(UPSpep25)
obj <- wrapper.mvImputation(UPSpep25, "QRILC")
condition1 <- '25fmol'
condition2 <- '10fmol'
qData <- Biobase::exprs(obj)
samplesData <- Biobase::pData(obj)
labels <- Biobase::pData(obj)[,"Label"]
limma <- diffAnaLimma(qData,samplesData, labels, condition1, condition2)
diffAnaComputeFDR(limma)
}
\author{
Samuel Wieczorek
}
