% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaComputeFDR}
\alias{diffAnaComputeFDR}
\title{Computes the FDR corresponding to the p-values of the 
differential analysis using}
\usage{
diffAnaComputeFDR(FC, pval, threshold_PVal = 0, threshold_LogFC = 0,
  pi0Method = 1)
}
\arguments{
\item{FC}{The result of the differential analysis processed 
by \code{\link{wrapper.limmaCompleteTest}}}

\item{pval}{The result of the differential analysis processed 
by \code{\link{wrapper.limmaCompleteTest}}}

\item{threshold_PVal}{The threshold on p-pvalue to
distinguish between differential and non-differential data}

\item{threshold_LogFC}{The threshold on log(Fold Change) to
distinguish between differential and non-differential data}

\item{pi0Method}{The parameter pi0.method of the method adjust.p 
in the package \code{cp4p}}
}
\value{
The computed FDR value (floating number)
}
\description{
This function is a wrappper to the function adjust.p from the
cp4p package. It returns the FDR corresponding to the p-values of the 
differential analysis.
The FDR is computed with the function \code{p.adjust}\{stats\}..
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[1:1000]
lapala <- findLapalaBlock(obj)
obj <- wrapper.impute.detQuant(obj)
obj <- reIntroduceLapala(obj, lapala)
obj <- wrapper.impute.detQuant(obj)
limma <- wrapper.limmaCompleteTest(obj, 1)
diffAnaComputeFDR(limma$FC,limma$P_Value)
}
\author{
Samuel Wieczorek
}
