% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaGetSignificant}
\alias{diffAnaGetSignificant}
\title{Returns a MSnSet object with only proteins significant after 
differential analysis.}
\usage{
diffAnaGetSignificant(obj)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}
}
\value{
A MSnSet
}
\description{
Returns a MSnSet object with only proteins significant after 
differential analysis.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept
keepThat <- mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
qData <- Biobase::exprs(obj)
sTab <- Biobase::pData(obj)
limma <- limmaCompleteTest(qData,sTab)
fc <- limma$FC[1]
pval <- limma$P_Value[1]
params <- list(design="OnevsOne", method="limma", th_logFC=0)
obj <- diffAnaSave(obj, limma,list(FC=fc, P_Value = pval), params)
signif <- diffAnaGetSignificant(obj)
}
\author{
Alexia Dorffer
}
