% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaGetSignificant}
\alias{diffAnaGetSignificant}
\title{Returns a MSnSet object with only proteins significant after 
differential analysis.}
\usage{
diffAnaGetSignificant(obj)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}
}
\value{
A MSnSet
}
\description{
Returns a MSnSet object with only proteins significant after 
differential analysis.
}
\examples{
data(UPSprotx2)
condition1 <- "10fmol"
condition2 <- "5fmol"
resLimma <- wrapper.diffAnaLimma(UPSprotx2, condition1, condition2)
obj <-diffAnaSave(UPSprotx2, resLimma, "limma", condition1, condition2)
signif <- diffAnaGetSignificant(obj)
}
\author{
Alexia Dorffer
}

