% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaGetSignificant}
\alias{diffAnaGetSignificant}
\title{Returns a MSnSet object with only proteins significant after 
differential analysis.}
\usage{
diffAnaGetSignificant(obj)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}
}
\value{
A MSnSet
}
\description{
Returns a MSnSet object with only proteins significant after 
differential analysis.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept
lapala <- findLapalaBlock(obj)
obj <- wrapper.impute.detQuant(obj)
obj <- reIntroduceLapala(obj, lapala)
obj <- wrapper.impute.detQuant(obj)
limma <- wrapper.limmaCompleteTest(obj, 1)
fc <- limma$FC[1]
pval <- limma$P_Value[1]
obj <- diffAnaSave(obj, list(logFC=fc, P_Value = pval))
signif <- diffAnaGetSignificant(obj)
}
\author{
Alexia Dorffer
}
