% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaLimma}
\alias{diffAnaLimma}
\title{Performs differential analysis on
an MSnSet object, calling the \code{limma} package functions}
\usage{
diffAnaLimma(qData, samplesData, labels, condition1, condition2)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{samplesData}{A dataframe where lines correspond to samples and columns to the meta-data for those samples.}

\item{labels}{A vector of the conditions (labels) (one label per sample).}

\item{condition1}{A vector that contains the names of the conditions 
considered as condition 1}

\item{condition2}{A vector that contains the names of the conditions 
considered as condition 2}
}
\value{
A dataframe as returned by the \code{limma} package
}
\description{
Method to perform differential analysis on
an \code{\link{MSnSet}} object (calls the \code{limma} package function).
}
\examples{
data(UPSpep25)
condition1 <- '25fmol'
condition2 <- '10fmol'
qData <- exprs(UPSpep25)
samplesData <- pData(UPSpep25)
labels <- pData(UPSpep25)[,"Label"]
diffAnaLimma(qData, samplesData, labels, condition1, condition2)
}
\author{
Florence Combes, Samuel Wieczorek
}

