% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaSave}
\alias{diffAnaSave}
\title{Returns a \code{MSnSet} object with the results of
the differential analysis performed with \code{\link{limma}} package.}
\usage{
diffAnaSave(obj, data, method = "limma", threshold_pVal = 1e-60,
  threshold_logFC = 0, fdr = 0, calibrationMethod = "pounds")
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{data}{The result of the differential analysis processed 
by \code{\link{wrapper.limmaCompleteTest}}}

\item{method}{The method used for differential analysis. 
Available choices are : "limma", "Welch"}

\item{threshold_pVal}{A float that indicates the threshold on p-value 
choosen to discriminate differential proteins.}

\item{threshold_logFC}{A float that indicates the threshold on 
log(Fold Change) to discriminatedifferential proteins.}

\item{fdr}{The FDR based on the values of threshold_pVal and 
threshold_logFC}

\item{calibrationMethod}{The calibration method used to compute the 
calibration plot}
}
\value{
A MSnSet
}
\description{
This method returns a class \code{MSnSet} object with the results
of differential analysis.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept
lapala <- findLapalaBlock(obj)
obj <- wrapper.impute.detQuant(obj)
obj <- reIntroduceLapala(obj, lapala)
obj <- wrapper.impute.detQuant(obj)
limma <- wrapper.limmaCompleteTest(obj, 1)
fc <- limma$FC[1]
pval <- limma$P_Value[1]
diffAnaSave(obj, list(logFC=fc, P_Value = pval))
}
\author{
Alexia Dorffer, Samuel Wieczorek
}
