% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaWelch}
\alias{diffAnaWelch}
\title{Performs a differential analysis on a \code{\link{MSnSet}} object
using the Welch t-test}
\usage{
diffAnaWelch(qData, labels, condition1, condition2)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{labels}{A vector of the conditions (labels) (one label per sample).}

\item{condition1}{A vector containing the names of the conditions 
qData as condition 1}

\item{condition2}{A vector containing the names of the conditions 
considered as condition 2}
}
\value{
A dataframe with two slots : P.Value (for the p-value) and logFC
(the log of the Fold Change).
}
\description{
Computes differential analysis on
an \code{\link{MSnSet}} object, using the Welch t-test
(\code{\link{t.test}{stats}}).
}
\examples{
data(UPSpep25)
condition1 <- '25fmol'
condition2 <- '10fmol'
qData <- Biobase::exprs(UPSpep25)
labels <- Biobase::pData(UPSpep25)[,"Label"]
diffAnaWelch(qData, labels, condition1, condition2)
}
\author{
Florence Combes, Samuel Wieczorek
}
