% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{finalizeAggregation}
\alias{finalizeAggregation}
\title{Finalizes the aggregation process}
\usage{
finalizeAggregation(obj.pep, pepData, protData, X)
}
\arguments{
\item{obj.pep}{A peptide object of class \code{MSnset}}

\item{pepData}{xxxx}

\item{protData}{xxxxx}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}
}
\value{
A protein object of class \code{MSnset}
}
\description{
Method to finalize the aggregation process
}
\author{
Samuel Wieczorek
}
