% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{formatPHResults}
\alias{formatPHResults}
\title{Extract logFC and raw pvalues from multiple post-hoc models summaries}
\usage{
formatPHResults(post_hoc_models_summaries)
}
\arguments{
\item{post_hoc_models_summaries}{a list of summaries of post-hoc models.}
}
\value{
a list of 2 dataframes containing the logFC values and pvalues for
each comparison.
}
\description{
Extract logFC and raw pvalues from multiple post-hoc models summaries
}
\examples{
utils::data(Exp1_R25_prot, package='DAPARdata')
obj <- Exp1_R25_prot[1:1000]
keepThat <- mvFilterGetIndices(obj, condition = "WholeMatrix", threshold=ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
anova_tests <- t(apply(Biobase::exprs(obj),1, classic1wayAnova, conditions=as.factor(Biobase::pData(obj)$Condition)))
names(anova_tests) <- rownames(Biobase::exprs(obj))
tms <- lapply(anova_tests,
             function(x) summary(multcomp::glht(x, linfct = multcomp::mcp(conditions = "Tukey")),
                   test = multcomp::adjusted("none")))
res <- formatPHResults(tms)

}
\author{
Hélène Borges
}
