% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{formatPHResults}
\alias{formatPHResults}
\title{Extract logFC and raw pvalues from multiple post-hoc models summaries}
\usage{
formatPHResults(post_hoc_models_summaries)
}
\arguments{
\item{post_hoc_models_summaries}{a list of summaries of post-hoc models.}
}
\value{
a list of 2 dataframes containing the logFC values and pvalues for
each comparison.
}
\description{
Extract logFC and raw pvalues from multiple post-hoc models summaries
}
\examples{
utils::data(Exp1_R25_prot, package='DAPARdata')
obj <- Exp1_R25_prot[1:1000]
level <- obj@experimentData@other$typeOfData
metacell.mask <- match.metacell(GetMetacell(obj), 'missing', level)
indices <- GetIndices_WholeMatrix(metacell.mask, op='>=', th=1)
obj <- MetaCellFiltering(obj, indices, cmd='delete')
anova_tests <- t(apply(exprs(obj$new),1, classic1wayAnova, 
conditions=as.factor(pData(obj$new)$Condition)))
names(anova_tests) <- rownames(exprs(obj$new))
tms <- lapply(anova_tests,
             function(x) summary(multcomp::glht(x, 
             linfct = multcomp::mcp(conditions = "Tukey")),
                   test = multcomp::adjusted("none")))
res <- formatPHResults(tms)

}
\author{
Hélène Borges
}
