% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getIndicesConditions}
\alias{getIndicesConditions}
\title{Gets the conditions indices}
\usage{
getIndicesConditions(obj, cond1, cond2)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{cond1}{A vector of Labels (a slot in the \code{pData()} table) for
the condition 1.}

\item{cond2}{A vector of Labels (a slot in the \code{pData()} table) for
the condition 2.}
}
\value{
A list with two slots \code{iCond1} and \code{iCond2} containing
respectively the indices of samples in the \code{pData()} table of the
dataset.
}
\description{
Returns a list for the two conditions where each slot is a vector of indices
for the samples in the \code{pData()} table
}
\examples{
data(UPSprotx2)
getIndicesConditions(UPSprotx2, "Cut3", "WT")
}
\author{
Florence Combes, Samuel Wieczorek
}

