% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNumberOfEmptyLines}
\alias{getNumberOfEmptyLines}
\title{Returns the number of empty lines in the data}
\usage{
getNumberOfEmptyLines(qData)
}
\arguments{
\item{qData}{A matrix corresponding to the quantitative data.}
}
\value{
An integer
}
\description{
Returns the number of empty lines in a matrix.
}
\examples{
data(UPSpep25)
qData <- Biobase::exprs(UPSpep25)
getNumberOfEmptyLines(qData)
}
\author{
Samuel Wieczorek
}
