% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPaletteForLabels}
\alias{getPaletteForLabels}
\title{Palette for plots in \code{DAPAR}}
\usage{
getPaletteForLabels(labels)
}
\arguments{
\item{labels}{A vector of labels (strings).}
}
\value{
A palette designed for the data manipulated in \code{DAPAR}
}
\description{
Selects colors for the plots in \code{DAPAR} based on the different conditions in
the dataset. The palette is derived from
the brewer palette "Dark2" (see \code{\link{RColorBrewer}}).
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
labels <- Biobase::pData(Exp1_R25_pept)[,"Label"]
getPaletteForLabels(labels)
}
\author{
Florence Combes, Samuel Wieczorek
}
