% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPaletteForLabels}
\alias{getPaletteForLabels}
\title{Palette for plots in DAPAR}
\usage{
getPaletteForLabels(labels)
}
\arguments{
\item{labels}{A vector of labels (strings).}
}
\value{
A palette designed for the data manipulated in DAPAR
}
\description{
Selects colors for the plots in DAPAR based on the different conditions in
the dataset. The palette is derived from
the brewer palette "Dark2" (see \code{\link{RColorBrewer}}).
}
\examples{
data(UPSpep25)
labels <- pData(UPSpep25)[,"Label"]
getPaletteForLabels(labels)
}
\author{
Florence Combes, Samuel Wieczorek
}

