% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPaletteForReplicates_HC}
\alias{getPaletteForReplicates_HC}
\title{Palette for highcharter plot the replicates in DAPAR}
\usage{
getPaletteForReplicates_HC(nColors)
}
\arguments{
\item{nColors}{The desired number of colors}
}
\value{
A palette designed for the data manipulated in \code{DAPAR}
}
\description{
Selects colors for the highcharter plots in \code{DAPAR} based on the replicates in
the dataset. The palette is derived from the brewer palette "Dark2" (see \code{\link{RColorBrewer}}).
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
n <- nrow(Biobase::pData(Exp1_R25_pept))
getPaletteForReplicates_HC(n)
}
\author{
Samuel Wieczorek
}
