% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getProcessingInfo}
\alias{getProcessingInfo}
\title{Returns the contains of the slot processing  of an object of 
class \code{\link{MSnSet}}}
\usage{
getProcessingInfo(obj)
}
\arguments{
\item{obj}{An object (peptides) of class \code{\link{MSnSet}}.}
}
\value{
The slot processing of obj@processingData
}
\description{
Returns the contents of the slot processing of an object of class \code{\link{MSnSet}}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
getProcessingInfo(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
