% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{getProteinsStats}
\alias{getProteinsStats}
\title{computes the number of proteins that are only defined by 
specific peptides, shared peptides or a mixture of two.}
\usage{
getProteinsStats(matUnique, matShared)
}
\arguments{
\item{matUnique}{The adjacency matrix with only specific peptides.}

\item{matShared}{The adjacency matrix with both specific and shared peptides.}
}
\value{
A list
}
\description{
This function computes the number of proteins that are only defined by 
specific peptides, shared peptides or a mixture of two.
}
\examples{
data(UPSpep25)
protID <- "Protein.group.IDs"
MShared <- BuildAdjacencyMatrix(UPSpep25, protID, FALSE)
MUnique <- BuildAdjacencyMatrix(UPSpep25, protID, TRUE)
getProteinsStats(MUnique,MShared)
}
\author{
Samuel Wieczorek
}

