% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{globalAdjPval}
\alias{globalAdjPval}
\title{xxx}
\usage{
globalAdjPval(x, pval.threshold = 1.05, method = 1, display = T)
}
\arguments{
\item{x}{xxx}

\item{pval.threshold}{xxx}

\item{method}{xxx}

\item{display}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
NULL

}
\author{
Thomas Burger
}
