% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{GlobalQuantileAlignment}
\alias{GlobalQuantileAlignment}
\title{Normalisation GlobalQuantileAlignement}
\usage{
GlobalQuantileAlignment(qData)
}
\arguments{
\item{qData}{xxxx}
}
\value{
A normalized numeric matrix
}
\description{
Normalisation GlobalQuantileAlignement
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
qData <- Biobase::exprs(Exp1_R25_pept)
normalized <- GlobalQuantileAlignment(qData)

}
\author{
Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier,
Enora Fremy
}
