% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{impute.detQuant}
\alias{impute.detQuant}
\title{xxxxxxxxx}
\usage{
impute.detQuant(qData, values)
}
\arguments{
\item{qData}{xxxxxxxxx}

\item{values}{xxxxxxxxx}
}
\value{
xxxxxxxxx
}
\description{
This method is xxxxxxxx
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qData <- Biobase::exprs(Exp1_R25_pept)
values <- getQuantile4Imp(qData)$shiftedImpVal
impute.detQuant(qData, values) 
}
\author{
Thomas Burger
}
