% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.iter}
\alias{inner.aggregate.iter}
\title{xxxx}
\usage{
inner.aggregate.iter(qData, X, ..., init.method, method)
}
\arguments{
\item{qData}{xxxxx}

\item{X}{xxxx}

\item{...}{xxxx}

\item{init.method}{xxx}

\item{method}{xxx}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\author{
Samuel Wieczorek
}
