% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.iter}
\alias{inner.aggregate.iter}
\title{xxxx}
\usage{
inner.aggregate.iter(
  pepData,
  X,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)
}
\arguments{
\item{pepData}{xxxxx}

\item{X}{xxxx}

\item{init.method}{xxx}

\item{method}{xxx}

\item{n}{xxxx}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\examples{
data(Exp1_R25_pept)
obj <- Exp1_R25_pept
protID <- "Protein_group_IDs"
X <- BuildAdjacencyMatrix(obj[seq_len(10)], protID, FALSE)
qdata.agg <- DAPAR::inner.aggregate.iter(Biobase::exprs(obj[seq_len(10)]), X)

}
