% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.topn}
\alias{inner.aggregate.topn}
\title{xxxx}
\usage{
inner.aggregate.topn(qData, X, n, method = "mean")
}
\arguments{
\item{qData}{xxxxx}

\item{X}{xxxx}

\item{n}{xxxxx}

\item{method}{xxxxx}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\author{
Samuel Wieczorek
}
