% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{limmaCompleteTest}
\alias{limmaCompleteTest}
\title{Computes a hierarchical differential analysis}
\usage{
limmaCompleteTest(qData, Conditions, RepBio, RepTech, Contrast = 1)
}
\arguments{
\item{qData}{A matrix of quantitative data, without any missing values.}

\item{Conditions}{A vector of factor which indicates the name of the biological condition for each replicate.}

\item{RepBio}{A vector of factor which indicates the number of the bio rep for each replicate.}

\item{RepTech}{A vector of factor which indicates the number of the tech rep for each replicate.}

\item{Contrast}{Indicates if the test concists of the comparison of each biological condition versus 
each of the other ones (Contrast=1; for example H0:"C1=C2" vs H1:"C1!=C2", etc.) 
or each condition versus all others (Contrast=2; e.g.  H0:"C1=(C2+C3)/2" vs
 H1:"C1!=(C2+C3)/2", etc. if there are three conditions).}
}
\value{
fdsfdgfdg
}
\description{
This function is a limmaCompleteTest
}
\examples{
data(UPSpep25)
obj <- wrapper.mvImputation(UPSpep25, "QRILC")
condition1 <- '25fmol'
condition2 <- '10fmol'
qData <- Biobase::exprs(obj)
RepBio <- RepTech <- factor(1:6)
conds <- factor(c(rep(condition1, 3), (rep(condition2, 3))))
limma <- limmaCompleteTest(qData,conds,RepBio, RepTech)
}
\author{
Quentin Giai-Gianetto
}

