% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{match.metacell}
\alias{match.metacell}
\title{Similar to the function \code{is.na} but focused on the equality with the paramter 'type'.}
\usage{
match.metacell(data, type, level = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{type}{The value to search in the dataframe}

\item{level}{xxx}
}
\value{
A boolean dataframe
}
\description{
Similar to the function \code{is.na} but focused on the equality with the paramter 'type'.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept[1:10,]
data <- Biobase::fData(obj)[, obj@experimentData@other$names_metacell]
match.metacell(data, type="MV_MEC", level = 'peptide')

}
\author{
Samuel Wieczorek
}
