% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{MeanPeptides}
\alias{MeanPeptides}
\title{Compute the intensity of proteins as the mean of the intensities
of their peptides.}
\usage{
MeanPeptides(matAdj, expr)
}
\arguments{
\item{matAdj}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{expr}{A matrix of intensities of peptides}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins as the mean of the 
intensities of their peptides.
}
\examples{
\dontrun{
data(UPSpepx2)
protID <- "Protein.group.IDs"
matAdj <- BuildAdjacencyMatrix(UPSpepx2, protID, FALSE)
meanPeptides(matAdj, exprs(UPSpepx2))
}
}
\author{
Alexia Dorffer
}

