% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{metacell.def}
\alias{metacell.def}
\title{Metadata vocabulary for entities}
\usage{
metacell.def(level = NULL)
}
\arguments{
\item{level}{A string designing the type of entity/pipeline. Available values are:
`peptide`, `protein`}
}
\description{
This function gives the vocabulary used for the metadata of each entity in
each condition.
Peptide-level vocabulary

├── 1.0 Quantitative Value
|    |
│    |── 1.1 Direct
|    |
│    |── 1.2 Indirect
│
├── 2.0 Missing value
|    |
│    |── 2.1 Missing POV
|    |
│    |── 2.2 Missing MEC
│
├── 3.0 Imputed value
|    |
│    |── 3.1 Imputed POV
|    |
│    |── 3.2 Imputed MEC
       
Protein-level vocabulary: same as peptide-level with one more category (Combined Value)
}
\details{
#' @title xxx
#' 
#' @description
#' xxxx
#' 
#' @export
#' 
controled.vocable <- function(){
  list( 'direct' =           'quantiValue-direct',
        'indirect' =         'quantiValue-indirect',
        'missingValue' =     'missingValue',
        'NA' =               'missingValue-NA',
        'POV' =              'missingValue-NA-POV',
        'POV_MCAR' =         'missingValue-NA-POV-MCAR',
        'POV_MNAR' =         'missingValue-NA-POV-MNAR',
        'MEC' =              'missingValue-NA-MEC',
        'MEC_MCAR' =         'missingValue-NA-MEC-MCAR',
        'MEC_MNAR' =         'missingValue-NA-MEC-MNAR',
        'imputed' =          'missingValue-imputed',
        'imputed_algo' =     'missingValue-imputed-algo',
        'unknown' =          'unknown') 

}
}
\author{
Thomas Burger, Samuel Wieczorek
}
