% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{metacell.def}
\alias{metacell.def}
\title{Metadata vocabulary for entities}
\usage{
metacell.def(level = NULL)
}
\arguments{
\item{level}{A string designing the type of entity/pipeline. Available values are:
`peptide`, `protein`}
}
\description{
This function gives the vocabulary used for the metadata of each entity in
each condition.
Peptide-level vocabulary
   1. Quantitative Value (= soit non defini par l'utilisateur, mais valeur de quanti présente (c'est donc un genre de "undefined"); soit les sous-catégories ci-dessous)
       1.1. Direct (=byMS/MS)
       1.2. Indirect (=bymatching)
   2. Missing value (= statut pour toutes les cellules, dont la valeur était manquantes à la création du dataset)
       2.1. Missing POV (=après diagnostic du nb de MV par condition)
       2.2. Missing MEC (=après diagnostic du nb de MV par condition)
   3. Imputed value
       3.1. Imputed POV (=2.1 ayant subi une imputation)
       3.2. Imputed MEC (=2.2 ayant subi une imputation)
       
Protein-level vocabulary: same as peptide-level with one more category (Combined Value)
}
\author{
Thomas Burger, Samuel Wieczorek
}
