% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{Metacell_maxquant}
\alias{Metacell_maxquant}
\title{Sets the metacell dataframe}
\usage{
Metacell_maxquant(qData, conds, df, level = NULL)
}
\arguments{
\item{qData}{An object of class \code{MSnSet}}

\item{conds}{xxx}

\item{df}{A list of integer xxxxxxx}

\item{level}{xxx}
}
\value{
xxxxx
}
\description{
Conversion rules for maxquant
Quanti         Identification       								            Tag
 == 0			    whatever ('By MS/MS', N.A., 'By Matching')	    NA
 > 0				    'By MS/MS'											                direct
 > 0				    'By matching'										                indirect
}
\author{
Samuel Wieczorek
}
