% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{MetacellFilteringScope}
\alias{MetacellFilteringScope}
\title{Lists the metacell scopes for filtering}
\usage{
MetacellFilteringScope()
}
\description{
Lists the metacell scopes for filtering
}
