% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell_Plots.R
\name{metacellPerLinesHisto_HC}
\alias{metacellPerLinesHisto_HC}
\title{Bar plot of missing values per lines using highcharter}
\usage{
metacellPerLinesHisto_HC(
  obj,
  pattern,
  detailed = FALSE,
  indLegend = "auto",
  showValues = FALSE
)
}
\arguments{
\item{obj}{xxx.}

\item{pattern}{xxx}

\item{detailed}{'value' or 'percent'}

\item{indLegend}{The indice of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates whether numeric values should be
drawn above the bars.}
}
\value{
A bar plot
}
\description{
This method plots a bar plot which represents the distribution of the 
number of missing values (NA) per lines (ie proteins).
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept[1:10,]
metacellPerLinesHisto_HC(obj, pattern = 'missing')

}
\author{
Florence Combes, Samuel Wieczorek
}
