% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{mvHisto}
\alias{mvHisto}
\title{Histogram of missing values}
\usage{
mvHisto(obj, indLegend = "auto", showValues = FALSE)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{indLegend}{The indices of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A histogram
}
\description{
This method plots a histogram of missing values.
}
\examples{
data(UPSprotx2)
mvHisto(UPSprotx2,showValues=TRUE)
}
\author{
Florence Combes, Samuel Wieczorek
}

