% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_mv_histo.R
\name{mvHisto_HC}
\alias{mvHisto_HC}
\title{Histogram of missing values}
\usage{
mvHisto_HC(
  qData,
  samplesData,
  conds,
  indLegend = "auto",
  showValues = FALSE,
  palette = NULL
)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{samplesData}{A dataframe where lines correspond to samples and 
columns to the meta-data for those samples.}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{indLegend}{The indices of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}

\item{palette}{xxx}
}
\value{
A histogram
}
\description{
#' This method plots a histogram of missing values.
#' 
#' @title Histogram of missing values
#' @param qData A dataframe that contains quantitative data.
#' @param samplesData A dataframe where lines correspond to samples and 
#' columns to the meta-data for those samples.
#' @param conds A vector of the conditions (one condition per sample).
#' @param indLegend The indices of the column name's in \code{pData()} tab
#' @param showValues A logical that indicates wether numeric values should be
#' drawn above the bars.
#' @param palette xxx
#' @return A histogram
#' @author Florence Combes, Samuel Wieczorek
#' @examples
#' utils::data(Exp1_R25_pept, package='DAPARdata')
#' qData <- Biobase::exprs(Exp1_R25_pept)
#' samplesData <- Biobase::pData(Exp1_R25_pept)
#' conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
#' mvHisto(qData, samplesData, conds, indLegend="auto", showValues=TRUE)
#' @export
mvHisto <- function(qData, samplesData, conds, indLegend="auto", showValues=FALSE, palette=NULL){
  
  if (is.null(palette)){
    palette <- RColorBrewer::brewer.pal(length(unique(conds)),"Dark2")
  }else{
    if (length(palette) != ncol(qData)){
      warning("The color palette has not the same dimension as the number of samples")
      return(NULL)
    }
  }
  
  if (identical(indLegend,"auto")) { 
    indLegend <- c(2:length(colnames(samplesData)))
  }
  
  
  colnames(qData) <- samplesData[,"Condition"]
  
  coeffMax <- .1
  
  NbNAPerCol <- colSums(is.na(qData))
  NbNAPerRow <- rowSums(is.na(qData))
  
  if (sum(NbNAPerCol) == 0) {if (sum(NbNAPerCol) == 0){
    NbNAPerCol <- rep(0,1+ncol(qData))
  }} 
  x <- barplot(NbNAPerCol, 
               main = "# NA per columns",
               col=palette,
               las=1,
               ylim = c(0, 1.2*max(1,NbNAPerCol)),
               names.arg = c(1:18), 
               cex.names=1.5,
               cex.axis=1.5,
               axisnames = FALSE
  )
  
  par(xpd = TRUE)
  graphics::text(x, -3,
                 label = colnames(qData),
                 srt = 45,
                 adj=1,
                 cex=1.4)
  
}
This method plots a histogram of missing values. Same as the function \code{mvHisto}
but uses the package \code{highcharter}
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
qData <- Biobase::exprs(Exp1_R25_pept)
samplesData <- Biobase::pData(Exp1_R25_pept)
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
mvHisto_HC(qData, samplesData, conds, indLegend="auto", showValues=TRUE)
}
\author{
Florence Combes, Samuel Wieczorek
}
