% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{mvImputation}
\alias{mvImputation}
\title{Missing values imputation from a matrix}
\usage{
mvImputation(qData, method)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{method}{The imputation method to be used.
Choices are QRILC, KNN, BPCA and MLE.}
}
\value{
The matrix imputed
}
\description{
This method is a wrapper to the \code{imputeLCMD} package adapted to a matrix.
}
\examples{
library(DAPARdata)
data(UPSprot25)
qData <- exprs(UPSprot25)
mvImputation(qData, "QRILC")
}
\author{
Samuel Wieczorek
}

