% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeD}
\alias{normalizeD}
\title{Normalisation}
\usage{
normalizeD(qData, labels, family, method)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{labels}{A vector of strings containing the column "Label" of the \code{pData()}.}

\item{family}{One of the following : Global Rescaling, 
Median Centering, Mean Centering, Mean Centering Scaling.}

\item{method}{"Overall" or "within conditions".}
}
\value{
A matrix normalized
}
\description{
Provides several methods to normalize data from
a matrix.
They are organized in four main families : Strong Rescaling, 
Median Centering, Mean Centering, Mean CenteringScaling.
For the first family, two sub-categories are available : the sum by columns
and the quantiles method.
For the three other families, two categories are available : 
"Overall" which means that the value for each protein 
(ie line in the expression data tab) is computed over all the samples ;
"within conditions" which means that the value for each protein 
(ie line in the matrix) is computed condition
by condition.
}
\examples{
library(DAPARdata)
data(UPSpep25)
qData <- exprs(UPSpep25)
labels <- pData(UPSpep25)[,"Label"]
normalizeD(qData, labels, "Median Centering", "within conditions")
}
\author{
Florence Combes, Samuel Wieczorek
}

