% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-pipeline-class.R
\docType{class}
\name{PipelinePeptide-class}
\alias{PipelinePeptide-class}
\alias{.PipelinePeptide}
\title{PipelineTemplate class}
\arguments{
\item{...}{Additional arguments for supporting functions. See details.}
}
\value{
A \code{PipelineTemplate} object
}
\description{
The \code{PipelineTemplate} class inherits from the \code{PipelineTemplate} and serves as a template for
instanciate pipeline classes such as proteinPipeline, proteinPipeline, etc. It can be used to manage results of
diverse assays on a collection of specimen. Currently,  the class can handle
assays that are organized instances of
\code{\linkS4class{MSnSet}},
\code{matrix}.
}
\section{Slots}{

\describe{
\item{\code{analysis}}{A character vector that is the name of the MS study}

\item{\code{pipelineType}}{A character vector that indicates the type of data that are managed in this instance}

\item{\code{processes}}{xxx}
}}

\examples{
example("PipelineTemp")
library(DAPARdata)
data('Exp1_R25_pept')
obj <- Exp1_R25_pept
samples <- Biobase::pData(obj)
mae <- PipelinePeptide(analysis= 'test',pipelineType = 'peptide', dataType = 'peptide',
processes='original',experiments=list(original=obj), 
colData=samples)
}
