% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgs.require}
\alias{pkgs.require}
\title{Loads packages}
\usage{
pkgs.require(ll.deps)
}
\arguments{
\item{ll.deps}{A `character()` vector which contains packages names}
}
\description{
Checks if a package is available to load it
}
\examples{
pkgs.require('DAPAR')

}
\author{
Samuel Wieczorek
}
