% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{plotJitter_rCharts}
\alias{plotJitter_rCharts}
\title{Display a a jitter plot for CC}
\usage{
plotJitter_rCharts(df, clickFunction = NULL)
}
\arguments{
\item{df}{xxxx}

\item{clickFunction}{xxxx}
}
\value{
A plot
}
\description{
Display a a jitter plot for CC
}
\examples{
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[seq_len(100)]
X <- BuildAdjacencyMatrix(obj, "Protein_group_IDs", TRUE)
ll <- get.pep.prot.cc(X)
plotJitter_rCharts(ll)

}
\author{
Thomas Burger, Samuel Wieczorek
}
