% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{removeLines}
\alias{removeLines}
\title{Removes lines in the dataset based on a prefix string.}
\usage{
removeLines(obj, idLine2Delete = NULL, prefix = NULL)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{idLine2Delete}{The name of the column that correspond to the data to filter}

\item{prefix}{A character string that is the prefix to find in the data}
}
\value{
An object of class \code{\link{MSnSet}}.
}
\description{
This function removes lines in the dataset based on a prefix string.
}
\examples{
data(UPSpepx2)
removeLines(UPSpepx2, "Contaminant")
removeLines(UPSpepx2, "Reverse")
}
\author{
Samuel Wieczorek
}

