% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioAnalysis.R
\name{scatterplotEnrichGO_HC}
\alias{scatterplotEnrichGO_HC}
\title{A dotplot that shows the result of a GO enrichment, using the package \code{\link{highcharter}}}
\usage{
scatterplotEnrichGO_HC(ego, maxRes = 10, title = NULL)
}
\arguments{
\item{ego}{The result of the GO enrichment, provides either by the function
enrichGO in \code{\link{DAPAR}} or bye the function \code{enrichGO} of the packaage \code{\link{clusterProfiler}}
maxRes An integer which is the maximum number of categories to display in the plot}

\item{maxRes}{The maximum number of categories to display in the plot}

\item{title}{The title of the plot}
}
\value{
A dotplot
}
\description{
A scatter plot of GO enrichment analysis
}
\author{
Samuel Wieczorek
}
