% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{search.metacell.tags}
\alias{search.metacell.tags}
\title{xxxx}
\usage{
search.metacell.tags(pattern, level)
}
\arguments{
\item{pattern}{xxx}

\item{level}{xxx}
}
\description{
Gives tags containing pattern as parent and all its children
}
\examples{
search.metacell.tags('POV', 'peptide')
search.metacell.tags('MV_POV', 'peptide')
search.metacell.tags('quanti', 'peptide')

}
\author{
Samuel Wieczorek
}
