% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{SumPeptides}
\alias{SumPeptides}
\title{Compute the intensity of proteins with the sum of the intensities
of their peptides.}
\usage{
SumPeptides(matAdj, expr)
}
\arguments{
\item{matAdj}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{expr}{A matrix of intensities of peptides}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins based on the sum of the 
intensities of their peptides.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
protID <- "Protein.group.IDs"
M <- BuildAdjacencyMatrix(Exp1_R25_pept, protID, FALSE)
SumPeptides(M, Biobase::exprs(Exp1_R25_pept))
}
\author{
Alexia Dorffer
}
