% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{TopnPeptides}
\alias{TopnPeptides}
\title{Compute the intensity of proteins as the sum of the 
intensities of their n best peptides.}
\usage{
TopnPeptides(matAdj, expr, n)
}
\arguments{
\item{matAdj}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{expr}{A matrix of intensities of peptides}

\item{n}{The maximum number of peptides used to aggregate a protein.}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins as the sum of the 
intensities of their n best peptides.
}
\examples{
\dontrun{
library(DAPARdata)
data(UPSpepx2)
protID <- "Protein.group.IDs"
matAdj <- BuildAdjacencyMatrix(UPSpepx2, protID, FALSE)
topnPeptides(matAdj, exprs(UPSpepx2), 3)
}
}
\author{
Alexia Dorffer
}

