% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{UpdateMetacell}
\alias{UpdateMetacell}
\title{Update metacell after imputation}
\usage{
UpdateMetacell(obj, method = "", na.type)
}
\arguments{
\item{obj}{xxx}

\item{method}{xxx}

\item{na.type}{xxx}
}
\value{
xxx
}
\description{
Update the metacell information of missing values that were imputed
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
