% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UpdateMetacell}
\alias{UpdateMetacell}
\title{Update metacell after imputation}
\usage{
UpdateMetacell(obj = NULL, na.type = NULL, method = "")
}
\arguments{
\item{obj}{xxx}

\item{na.type}{xxx}

\item{method}{xxx}
}
\description{
Update the metacell information of missing values that were imputed
}
\author{
Samuel Wieczorek
}
