% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{varianceDistD}
\alias{varianceDistD}
\title{Distribution of variance of proteins}
\usage{
varianceDistD(qData, labels = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{labels}{A vector of the conditions (labels) (one label per sample).}
}
\value{
A density plot
}
\description{
Builds a densityplot of the variance of entities in the exprs() table
of a object. The variance is calculated for each condition (Label) present
in the dataset (see the slot \code{'Label'} in the \code{pData()} table)
}
\examples{
data(UPSpep25)
labels <- Biobase::pData(UPSpep25)[,"Label"]
varianceDistD(UPSpep25)
}
\author{
Florence Combes, Samuel Wieczorek
}
\seealso{
\code{\link{densityPlotD}}.
}

