% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_violin.R
\name{violinPlotD}
\alias{violinPlotD}
\title{Builds a violinplot from a dataframe}
\usage{
violinPlotD(
  obj,
  conds,
  keyId,
  legend = NULL,
  palette = NULL,
  subset.view = NULL
)
}
\arguments{
\item{obj}{xxx}

\item{conds}{xxx}

\item{keyId}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}

\item{subset.view}{xxx}
}
\value{
A violinplot
}
\description{
Builds a violinplot from a dataframe
}
\examples{
utils::data(Exp1_R25_prot, package='DAPARdata')
obj <- Exp1_R25_prot
library(vioplot)
legend <- conds <- Biobase::pData(obj)$Condition
key <- "Protein_IDs"
violinPlotD(obj, conds, key, legend, NULL, 1:10)
pal <- ExtendPalette(length(unique(conds)))
violinPlotD(obj, conds, key, legend, pal, 1:10)

}
\author{
Samuel Wieczorek, Anais Courtier
}
