% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_violin.R
\name{violinPlotD}
\alias{violinPlotD}
\title{Builds a violinplot from a dataframe}
\usage{
violinPlotD(
  qData,
  conds,
  keyId,
  legend = NULL,
  palette = NULL,
  subset.view = NULL
)
}
\arguments{
\item{qData}{xxx}

\item{conds}{xxx}

\item{keyId}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}

\item{subset.view}{xxx}
}
\value{
A violinplot
}
\description{
Builds a violinplot from a dataframe
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
library(vioplot)
legend <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
conds <- Biobase::pData(Exp1_R25_pept)[["Condition"]]
key <- "Protein_group_IDs"
violinPlotD(Biobase::exprs(Exp1_R25_pept), conds=legend, keyId=key, palette=c(rep('blue',3), rep('green',3)),subset.view=1:10)

}
\seealso{
\code{\link{densityPlotD}}
}
\author{
Samuel Wieczorek, Anais Courtier
}
