% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{vsn}
\alias{vsn}
\title{Normalisation vsn}
\usage{
vsn(qData, conds, type = NULL)
}
\arguments{
\item{qData}{A numeric matrix.}

\item{conds}{xxx}

\item{type}{"overall" (shift all the sample distributions at once) or
"within conditions" (shift the sample distributions within each condition
at a time).}
}
\value{
A normalized numeric matrix
}
\description{
Normalisation vsn
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
qData <- Biobase::exprs(Exp1_R25_pept)
conds <- Biobase::pData(Exp1_R25_pept)$Condition
normalized <- vsn(qData, conds, type = "overall")

}
\author{
Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy
}
