% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.boxPlotD}
\alias{wrapper.boxPlotD}
\title{Wrapper to the boxplotD function on an object \code{\link{MSnSet}}}
\usage{
wrapper.boxPlotD(obj, dataForXAxis = "Label", group2Color = "Condition")
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{dataForXAxis}{A vector of strings containing the names of columns 
in \code{pData()} to print labels on X-axis (Default is "Label").}

\item{group2Color}{A string that indicates how to color the replicates: one
color per condition (value "Condition") or one 
color per replicate (value "Replicate"). Default value is by Condition.}
}
\value{
A boxplot
}
\description{
This function is a wrapper for using the boxPlotD function with objects of 
class \code{\link{MSnSet}}
}
\examples{
data(UPSpep25)
types <- c("Label","Analyt.Rep")
wrapper.boxPlotD(UPSpep25, types)
}
\author{
Florence Combes, Samuel Wieczorek
}
\seealso{
\code{\link{wrapper.densityPlotD}}
}

