% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.boxPlotD_HC}
\alias{wrapper.boxPlotD_HC}
\title{Wrapper to the boxplotD_HC function on an object \code{MSnSet}}
\usage{
wrapper.boxPlotD_HC(obj, dataForXAxis = "Condition")
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{dataForXAxis}{A vector of strings containing the names of columns 
in \code{pData()} to print conditions on X-axis (Default is "Condition").}

\item{...}{xxx}
}
\value{
A boxplot
}
\description{
This function is a wrapper for using the boxPlotD_HC function with objects of 
class \code{MSnSet}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
types <- c("Condition","Bio.Rep")
wrapper.boxPlotD_HC(Exp1_R25_pept, types)
}
\seealso{
\code{\link{wrapper.densityPlotD}}
}
\author{
Samuel Wieczorek
}
