% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.corrMatrixD}
\alias{wrapper.corrMatrixD}
\title{Displays a correlation matrix of the quantitative data of the
\code{exprs()} table}
\usage{
wrapper.corrMatrixD(obj, rate = 5, indLegend = NULL)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{rate}{A float that defines the gradient of colors.}

\item{indLegend}{A vector of indices in the columns in
the \code{pData()} table choosen for the labels in the axes.}
}
\value{
A colored correlation matrix
}
\description{
Builds a correlation matrix based on a \code{\link{MSnSet}} object.
}
\examples{
data(UPSprotx2)
wrapper.corrMatrixD(UPSprotx2)
}
\author{
Alexia Dorffer
}

