% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_CV_dist.R
\name{wrapper.CVDistD_HC}
\alias{wrapper.CVDistD_HC}
\title{Distribution of CV of entities}
\usage{
wrapper.CVDistD_HC(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}}

\item{...}{arguments for palette.}
}
\value{
A density plot
}
\description{
#' Builds a densityplot of the CV of entities in the exprs() table
#' of an object \code{MSnSet}. The variance is calculated for each 
#' condition present
#' in the dataset (see the slot \code{'Condition'} in the \code{pData()} table).
#' 
#' @title Distribution of CV of entities
#' @param obj An object of class \code{MSnSet}.
#' @param ... arguments for palette
#' @return A density plot
#' @author Alexia Dorffer
#' @examples
#' utils::data(Exp1_R25_pept, package='DAPARdata')
#' wrapper.CVDistD(Exp1_R25_pept)
#' @importFrom Biobase exprs pData
#' @export
wrapper.CVDistD <- function(obj, ...){
  qData <- Biobase::exprs(obj)
  conds <- Biobase::pData(obj)[,"Condition"]
  DAPAR::CVDistD(qData, conds, ...)
}
Builds a densityplot of the CV of entities in the exprs() table. 
of an object \code{MSnSet}. The variance is calculated for each 
condition present
in the dataset (see the slot \code{'Condition'} in the \code{pData()} table).
Same as the function \code{\link{wrapper.CVDistD}} but uses the package \code{highcharter}
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
wrapper.CVDistD_HC(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
