% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.densityPlotD}
\alias{wrapper.densityPlotD}
\title{Builds a densityplot from an object of class \code{\link{MSnSet}}}
\usage{
wrapper.densityPlotD(obj, labelsForLegend = NULL, indData2Show = NULL,
  group2Color = "Condition")
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{labelsForLegend}{A vector of labels to show in densityplot.}

\item{indData2Show}{xxxxxxx
in the density plot.}

\item{group2Color}{xxxxx}
}
\value{
A density plot
}
\description{
This function is a wrapper for using the densityPlotD function with objects of class \code{\link{MSnSet}}
}
\examples{
data(UPSprotx2)
labels <- pData(UPSprotx2)[,"Label"]
wrapper.densityPlotD(UPSprotx2, labels)
}
\author{
Alexia Dorffer
}
\seealso{
\code{\link{wrapper.boxPlotD}}, \code{\link{wrapper.varianceDistD}}
}

