% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.densityPlotD}
\alias{wrapper.densityPlotD}
\title{Builds a densityplot from an object of class \code{\link{MSnSet}}}
\usage{
wrapper.densityPlotD(obj, labelsForLegend = NULL, indData2Show = NULL,
  group2Color = "Condition")
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{labelsForLegend}{A vector of labels to show in densityplot.}

\item{indData2Show}{A vector of the indices of the columns to show in 
the plot. The indices are those of indices of the columns int the data
frame qDataBefore in the density plot.}

\item{group2Color}{A string that indicates how to color the replicates: 
one color per condition (value "Condition") or one color per replicate
(value "Replicate"). Default value is by Condition.}
}
\value{
A density plot
}
\description{
This function is a wrapper for using the densityPlotD function with 
objects of class \code{\link{MSnSet}}
}
\examples{
data(UPSpep25)
labels <- pData(UPSpep25)[,"Label"]
wrapper.densityPlotD(UPSpep25, labels)
}
\author{
Alexia Dorffer
}
\seealso{
\code{\link{wrapper.boxPlotD}}, 
\code{\link{wrapper.varianceDistD}}
}

