% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{wrapper.diffAnaLimma}
\alias{wrapper.diffAnaLimma}
\title{Performs differential analysis on
an MSnSet object, calling the \code{limma} package functions}
\usage{
wrapper.diffAnaLimma(obj, condition1, condition2)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{condition1}{A vector that contains the names of the conditions 
considered as condition 1.}

\item{condition2}{A vector that contains the names of the conditions 
considered as condition 2.}
}
\value{
A dataframe as returned by the \code{limma} package
}
\description{
Method to perform differential analysis on
a \code{\link{MSnSet}} object (calls the \code{limma} package function).
}
\examples{
library(DAPARdata)
data(UPSprotx2)
condition1 <- '10fmol'
condition2 <- '5fmol'
wrapper.diffAnaLimma(UPSprotx2, condition1, condition2)
}
\author{
Alexia Dorffer
}

