% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.heatmapD}
\alias{wrapper.heatmapD}
\title{This function is a wrapper to \code{\link{heatmap.2}} that displays 
quantitative data in the \code{exprs()} table of an object of
class \code{\link{MSnSet}}}
\usage{
wrapper.heatmapD(obj, distance = "euclidean", cluster = "average",
  dendro = FALSE)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{distance}{The distance used by the clustering algorithm to compute 
the dendrogram. See \code{help(heatmap.2)}.}

\item{cluster}{the clustering algorithm used to build the dendrogram.
See \code{help(heatmap.2)}}

\item{dendro}{A boolean to indicate fi the dendrogram has to be displayed}
}
\value{
A heatmap
}
\description{
Builds a heatmap of the quantitative proteomic data of a \code{\link{MSnSet}} object.
}
\examples{
library(DAPAR)
data(testWithoutNA)
wrapper.heatmapD(testWithoutNA)
}
\author{
Alexia Dorffer
}

