% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{wrapper.impute.detQuant}
\alias{wrapper.impute.detQuant}
\title{Wrapper of the function \code{\link{impute.detQuant}} for objects
of class \code{MSnSet}}
\usage{
wrapper.impute.detQuant(obj = NULL, qval = 0.025, factor = 1, na.type = NULL)
}
\arguments{
\item{obj}{An instance of class \code{MSnSet}}

\item{qval}{An expression set containing quantitative values of various replicates}

\item{factor}{A scaling factor to multiply the imputation value with}

\item{na.type}{A string which indicates the type of missing values to impute. 
Available values are: `NA` (for both POV and MEC), `POV`, `MEC`.}
}
\value{
An imputed instance of class \code{MSnSet}
}
\description{
This method is a wrapper of the function \code{\link{impute.detQuant}} for objects
of class \code{MSnSet}
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
wrapper.impute.detQuant(Exp1_R25_pept)

}
\author{
Samuel Wieczorek
}
