% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{wrapper.impute.KNN}
\alias{wrapper.impute.KNN}
\title{KNN missing values imputation from a \code{MSnSet} object}
\usage{
wrapper.impute.KNN(obj = NULL, K, na.type)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{K}{the number of neighbors.}

\item{na.type}{A string which indicates the type of missing values to impute.
Available values are: `missing POV`.}
}
\value{
The object \code{obj} which has been imputed
}
\description{
Can impute only POV missing values. This method is a wrapper for
objects of class \code{MSnSet} and imputes missing values with a fixed value.
This function imputes the missing values condition by condition.
}
\examples{
data(Exp1_R25_pept)
obj.imp.pov <- wrapper.impute.KNN(
    obj = Exp1_R25_pept[seq_len(10)], K = 3,
    na.type = "missing POV"
)

}
\author{
Samuel Wieczorek
}
