% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{wrapper.impute.pa}
\alias{wrapper.impute.pa}
\title{Imputation of peptides having no values in a biological condition.}
\usage{
wrapper.impute.pa(obj)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}
}
\value{
The \code{exprs(obj)} matrix with imputed values instead of missing values.
}
\description{
This method is a wrapper to the function \code{impute.pa} of the package
\code{imp4p} adapted to an object of class \code{\link{MSnSet}}.
}
\examples{
data(UPSpep25)
dat <- mvFilter(UPSpep25, type="allCond", th = 1)
dat <- wrapper.impute.pa(dat)
}
\author{
Samuel Wieczorek
}
