% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{wrapper.impute.pa}
\alias{wrapper.impute.pa}
\title{Imputation of peptides having no values in a biological condition.}
\usage{
wrapper.impute.pa(obj = NULL, q.min = 0.025, na.type)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{q.min}{Same as the function \code{impute.pa} in the package \code{imp4p}}

\item{na.type}{A string which indicates the type of missing values to impute. 
Available values are: `NA` (for both POV and MEC).}
}
\value{
The \code{exprs(obj)} matrix with imputed values instead of missing values.
}
\description{
This method is a wrapper to the function \code{impute.pa} of the package
\code{imp4p} adapted to an object of class \code{MSnSet}.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept[1:10]
obj <- mvFilter(obj, type="AllCond", th = 1)
obj.imp.pov <- wrapper.impute.pa(obj, na.type = 'missing_POV')

}
\author{
Samuel Wieczorek
}
