% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{wrapper.impute.slsa}
\alias{wrapper.impute.slsa}
\title{Missing values imputation using the LSimpute algorithm.}
\usage{
wrapper.impute.slsa(obj, selec = 100)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{selec}{An integer which corresponds to the \code{selec} parameter of the
function \code{impute.slsa}.}
}
\value{
The \code{exprs(obj)} matrix with imputed values instead of missing values.
}
\description{
This method is a wrapper to the function \code{impute.slsa} of the package \code{imp4p} adapted to 
an object of class \code{\link{MSnSet}}.
}
\examples{
data(UPSpep25)
dat <- mvFilter(UPSpep25, type="allCond", th = 1)
wrapper.impute.slsa(dat)
}
\author{
Samuel Wieczorek
}

