% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation.R
\name{wrapper.mvImputation}
\alias{wrapper.mvImputation}
\title{Missing values imputation from a \code{\link{MSnSet}} object}
\usage{
wrapper.mvImputation(obj, method)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{method}{The imputation method to be used.
Choices are QRILC, KNN, BPCA and MLE.}
}
\value{
The object \code{obj} which has been imputed
}
\description{
This method is a wrapper to the \code{imputeLCMD} package adapted to 
objects of class \code{\link{MSnSet}}.
}
\examples{
library(DAPARdata)
data(UPSprotx2)
wrapper.mvImputation(UPSprotx2, "QRILC")
}
\author{
Alexia Dorffer
}

