% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.pca}
\alias{wrapper.pca}
\title{Compute the PCA}
\usage{
wrapper.pca(obj, var.scaling = TRUE, ncp = NULL)
}
\arguments{
\item{obj}{xxx}

\item{var.scaling}{The dimensions to plot}

\item{ncp}{xxxx}
}
\value{
A xxxxxx
}
\description{
Compute the PCA
}
\examples{
require(DAPARdata)
utils::data(Exp1_R25_pept)
 obj <- mvFilter(Exp1_R25_pept, "wholeMatrix", 6)
res.pca <- wrapper.pca(obj)
}
\author{
Samuel Wieczorek
}
