% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.t.tests.R
\name{wrapper.t_test_Complete}
\alias{wrapper.t_test_Complete}
\title{xxxxx}
\usage{
wrapper.t_test_Complete(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{MSnSet} with no missing values}

\item{...}{See \code{compute.t.tests}}
}
\value{
xxxxxxx
}
\description{
This function is a wrapper xxxxx
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[1:1000]
lapala <- findMECBlock(obj)
obj <- wrapper.impute.detQuant(obj)
obj <- reIntroduceMEC(obj, lapala)
obj <- wrapper.impute.detQuant(obj)
ttest <- wrapper.t_test_Complete(obj, 1)
}
\author{
Samuel Wieczorek
}
