% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.t.tests.R
\name{wrapper.t_test_Complete}
\alias{wrapper.t_test_Complete}
\title{xxxxx}
\usage{
wrapper.t_test_Complete(obj, design)
}
\arguments{
\item{obj}{An object of class \code{MSnSet} with no missing values}

\item{design}{An integer that reflects the type of comparisons. Available values are 1 (One vs One) or (One vs All)}
}
\value{
xxxxxxx
}
\description{
This function is a wrapper xxxxx
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[1:1000]
lapala <- findLapalaBlock(obj)
obj <- wrapper.impute.detQuant(obj)
obj <- reIntroduceLapala(obj, lapala)
obj <- wrapper.impute.detQuant(obj)
ttest <- wrapper.t_test_Complete(obj, 1)
}
\author{
Samuel Wieczorek
}
