% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.violinPlotD}
\alias{wrapper.violinPlotD}
\title{Wrapper to the violinPlotD function on an object \code{\link{MSnSet}}}
\usage{
wrapper.violinPlotD(obj, dataForXAxis = "Label", group2Color = "Condition")
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{dataForXAxis}{A vector of strings containing the names of columns 
in \code{pData()} to print labels on X-axis (Default is "Label").}

\item{group2Color}{A string that indicates how to color the replicates: one
color per condition (value "Condition") or one 
color per replicate (value "Replicate"). Default value is by Condition.}
}
\value{
A violin plot
}
\description{
This function is a wrapper for using the violinPlotD function with objects of 
class \code{\link{MSnSet}}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
library(vioplot)
types <- c("Label","Analyt.Rep")
wrapper.violinPlotD(Exp1_R25_pept, types)
}
\seealso{
\code{\link{wrapper.densityPlotD}}, \code{\link{wrapper.boxPlotD}}
}
\author{
Samuel Wieczorek
}
