% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{wrapperCalibrationPlot}
\alias{wrapperCalibrationPlot}
\title{Performs a calibration plot on
an \code{MSnSet} object, calling the \code{cp4p} package functions.}
\usage{
wrapperCalibrationPlot(vPVal, pi0Method = "pounds")
}
\arguments{
\item{vPVal}{A dataframe that contains quantitative data.}

\item{pi0Method}{A vector of the conditions (labels) (one label 
per sample).}
}
\value{
A plot
}
\description{
This function is a wrapper to the calibration.plot method of the 
\code{cp4p} package for use with \code{MSnSet} objects.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
condition1 <- '25fmol'
condition2 <- '10fmol'
qData <- Biobase::exprs(Exp1_R25_pept[1:1000])
labels <- Biobase::pData(Exp1_R25_pept[1:1000])[,"Label"]
diffAnaWelch(qData, labels, condition1, condition2)
}
\author{
Samuel Wieczorek
}
