% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{wrapperCalibrationPlot}
\alias{wrapperCalibrationPlot}
\title{xxxxxxxxxx}
\usage{
wrapperCalibrationPlot(vPVal, pi0Method = "pounds")
}
\arguments{
\item{vPVal}{A dataframe that contains quantitative data.}

\item{pi0Method}{A vector of the conditions (labels) (one label per sample).}
}
\value{
A plot
}
\description{
xxxxxxxxxxxxx
}
\examples{
data(UPSprotx2)
condition1 <- '10fmol'
condition2 <- '5fmol'
qData <- exprs(UPSprotx2)
labels <- pData(UPSprotx2)[,"Label"]
diffAnaWelch(qData, labels, condition1, condition2)
}
\author{
Samuel Wieczorek
}

