% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{writeMSnsetToExcel}
\alias{writeMSnsetToExcel}
\title{This function exports a \code{\link{MSnSet}} object to a Excel file.}
\usage{
writeMSnsetToExcel(obj, filename, id)
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}}.}

\item{filename}{A character string for the name of the Excel file.}

\item{id}{An integer to select in the fdata frame which column has to be
used as an index.}
}
\value{
A Excel file
}
\description{
This function exports a \code{\link{MSnSet}} data object to a Excel file.
Each of the
three data.frames in the \code{\link{MSnSet}} object (ie experimental data,
phenoData
and metaData are respectively integrated into separate sheets in
the Excel file).
}
\examples{
data(UPSprotx2)
writeMSnsetToExcel(UPSprotx2, "foo", 1)
}
\author{
Samuel Wieczorek
}

